fix the code and changed the ui
