+++ lib/models/azakdata.dart
@@ -1,0 +1,738 @@
+import 'package:flutter/material.dart';
+
+/// A simple data model for each dhikr/adhkar item
+class DhikrItem {
+  final String arabic;       // Full Arabic text
+  final String translation;  // We’ll store the short note e.g. "3 times" or "One time"
+  final int repeat;          // How many times
+
+  DhikrItem({
+    required this.arabic,
+    required this.translation,
+    required this.repeat,
+  });
+}
+
+/// 
+/// ──────────────────────────────────────────────────────────────────────────────
+/// SLEEP AZKAR
+/// ──────────────────────────────────────────────────────────────────────────────
+/// All items exactly as you provided (with "3times", "One time", "33 times", etc.)
+///
+final List<DhikrItem> sleepAzkar = [
+  DhikrItem(
+    arabic: '''بسم الله الرحمن الرحيم
+قُلۡ هُوَ ٱللَّهُ أَحَدٌ ١ ٱللَّهُ ٱلصَّمَدُ ٢ لَمۡ يَلِدۡ وَلَمۡ يُولَدۡ ٣ وَلَمۡ يَكُن لَّهُۥ كُفُوًا أَحَدُۢ ٤''',
+    translation: '3times',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic: '''بسم الله الرحمن الرحيم
+قُلۡ أَعُوذُ بِرَبِّ ٱلۡفَلَقِ ١ مِن شَرِّ مَا خَلَقَ ٢ وَمِن شَرِّ غَاسِقٍ إِذَا وَقَبَ ٣ وَمِن شَرِّ ٱلنَّفَّٰثَٰتِ فِي ٱلۡعُقَدِ ٤ وَمِن شَرِّ حَاسِدٍ إِذَا حَسَدَ ٥''',
+    translation: '3times',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic: '''بسم الله الرحمن الرحيم
+قُلۡ أَعُوذُ بِرَبِّ ٱلنَّاسِ ١ مَلِكِ ٱلنَّاسِ ٢ إِلَٰهِ ٱلنَّاسِ ٣ مِن شَرِّ ٱلۡوَسۡوَاسِ ٱلۡخَنَّاسِ ٤ ٱلَّذِي يُوَسۡوِسُ فِي صُدُورِ ٱلنَّاسِ ٥ مِنَ ٱلۡجِنَّةِ وَٱلنَّاسِ ٦''',
+    translation: '3times',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic: '''ٱللَّهُ لَآ إِلَٰهَ إِلَّا هُوَ ٱلۡحَيُّ ٱلۡقَيُّومُۚ لَا تَأۡخُذُهُۥ سِنَةٞ وَلَا نَوۡمٞۚ لَّهُۥ مَا فِي ٱلسَّمَٰوَٰتِ وَمَا فِي ٱلۡأَرۡضِۗ مَن ذَا ٱلَّذِي يَشۡفَعُ عِندَهُۥٓ إِلَّا بِإِذۡنِهِۦۚ يَعۡلَمُ مَا بَيۡنَ أَيۡدِيهِمۡ وَمَا خَلۡفَهُمۡۖ وَلَا يُحِيطُونَ بِشَيۡءٖ مِّنۡ عِلۡمِهِۦٓ إِلَّا بِمَا شَآءَۚ وَسِعَ كُرۡسِيُّهُ ٱلسَّمَٰوَٰتِ وَٱلۡأَرۡضَۖ وَلَا يَـُٔودُهُۥ حِفۡظُهُمَاۚ وَهُوَ ٱلۡعَلِيُّ ٱلۡعَظِيمُ ٢٥٥''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''ءَامَنَ ٱلرَّسُولُ بِمَآ أُنزِلَ إِلَيۡهِ مِن رَّبِّهِۦ وَٱلۡمُؤۡمِنُونَۚ كُلٌّ ءَامَنَ بِٱللَّهِ وَمَلَٰٓئِكَتِهِۦ وَكُتُبِهِۦ وَرُسُلِهِۦ لَا نُفَرِّقُ بَيۡنَ أَحَدٖ مِّن رُّسُلِهِۦۚ وَقَالُواْ سَمِعۡنَا وَأَطَعۡنَاۖ غُفۡرَانَكَ رَبَّنَا وَإِلَيۡكَ ٱلۡمَصِيرُ ٢٨٥
+لَا يُكَلِّفُ ٱللَّهُ نَفۡسًا إِلَّا وُسۡعَهَاۚ لَهَا مَا كَسَبَتۡ وَعَلَيۡهَا مَا ٱكۡتَسَبَتۡۗ رَبَّنَا لَا تُؤَاخِذۡنَآ إِن نَّسِينَآ أَوۡ أَخۡطَأۡنَاۚ رَبَّنَا وَلَا تَحۡمِلۡ عَلَيۡنَآ إِصۡرٗا كَمَا حَمَلۡتَهُۥ عَلَى ٱلَّذِينَ مِن قَبۡلِنَاۚ رَبَّنَا وَلَا تُحَمِّلۡنَا مَا لَا طَاقَةَ لَنَا بِهِۦۖ وَٱعۡفُ عَنَّا وَٱغۡفِرۡ لَنَا وَٱرۡحَمۡنَآۚ أَنتَ مَوۡلَىٰنَا فَٱنصُرۡنَا عَلَى ٱلۡقَوۡمِ ٱلۡكَٰفِرِينَ ٢٨٦''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''بِاسْمِكَ رَبِّي وَضَعْتُ جَنْبِي، وَبِكَ أَرْفَعُهُ، فَإِن أَمْسَكْتَ نَفْسِي فارْحَمْهَا، وَإِنْ أَرْسَلْتَهَا فَاحْفَظْهَا، بِمَا تَحْفَظُ بِهِ عِبَادَكَ الصَّالِحِينَ''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''اللَّهُمَّ إِنَّكَ خَلَقْتَ نَفْسِي وَأَنْتَ تَوَفَّاهَا، لَكَ مَمَاتُهَا وَمَحْياهَا، إِنْ أَحْيَيْتَهَا فَاحْفَظْهَا، وَإِنْ أَمَتَّهَا فَاغْفِرْ لَهَا. اللَّهُمَّ إِنِّي أَسْأَلُكَ العَافِيَةَ''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''اللَّهُمَّ قِنِي عَذَابَكَ يَوْمَ تَبْعَثُ عِبَادَكَ''',
+    translation: '3times',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic: '''بِاسْمِكَ اللَّهُمَّ أَمُوتُ وَأَحْيَا''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''سُبْحَانَ اللَّهِ (ثلاثاً وثلاثين) وَالْحَمْدُ لِلَّهِ (ثلاثاً وثلاثين) وَاللَّهُ أَكْبَرُ (أربعاً وثلاثينَ)''',
+    translation: '33 times',
+    repeat: 33,
+  ),
+  DhikrItem(
+    arabic: '''اللَّهُمَّ رَبَّ السَّمَوَاتِ السَّبْعِ وَرَبَّ الأَرْضِ، وَرَبَّ الْعَرْشِ الْعَظِيمِ، رَبَّنَا وَرَبَّ كُلِّ شَيْءٍ، فَالِقَ الْحَبِّ وَالنَّوَى، وَمُنْزِلَ التَّوْرَاةِ وَالْإِنْجِيلِ، وَالْفُرْقَانِ، أَعُوذُ بِكَ مِنْ شَرِّ كُلِّ شَيْءٍ أَنْتَ آخِذٌ بِنَاصِيَتِهِ. اللَّهُمَّ أَنْتَ الأَوَّلُ فَلَيْسَ قَبْلَكَ شَيْءٌ، وَأَنْتَ الآخِرُ فَلَيْسَ بَعْدَكَ شَيْءٌ، وَأَنْتَ الظَّاهِرُ فَلَيْسَ فَوْقَكَ شَيْءٌ، وَأَنْتَ الْبَاطِنُ فَلَيْسَ دُونَكَ شَيْءٌ، اقْضِ عَنَّا الدَّيْنَ وَأَغْنِنَا مِنَ الْفَقْرِ''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''الْحَمْدُ لِلَّهِ الَّذِي أَطْعَمَنَا وَسَقَانَا، وَكَفَانَا، وَآوَانَا، فَكَمْ مِمَّنْ لاَ كَافِيَ لَهُ وَلاَ مُؤْوِيَ''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''اللَّهُمَّ عَالِمَ الغَيْبِ وَالشَّهَادَةِ فَاطِرَ السَّمَوَاتِ وَالْأَرْضِ، رَبَّ كُلِّ شَيْءٍ وَمَلِيكَهُ، أَشْهَدُ أَنْ لاَ إِلَهَ إِلاَّ أَنْتَ، أَعُوذُ بِكَ مِنْ شَرِّ نَفْسِي، وَمِنْ شَرِّ الشَّيْطانِ وَشِرْكِهِ، وَأَنْ أَقْتَرِفَ عَلَى نَفْسِي سُوءاً، أَوْ أَجُرَّهُ إِلَى مُسْلِمٍ''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''بسم الله الرحمن الرحيم
+الٓمٓ ١ تَنزِيلُ ٱلۡكِتَٰبِ لَا رَيۡبَ فِيهِ مِن رَّبِّ ٱلۡعَٰلَمِينَ ٢ أَمۡ يَقُولُونَ ٱفۡتَرَىٰهُۚ بَلۡ هُوَ ٱلۡحَقُّ مِن رَّبِّكَ لِتُنذِرَ قَوۡمٗا مَّآ أَتَىٰهُم مِّن نَّذِيرٖ مِّن قَبۡلِكَ لَعَلَّهُمۡ يَهۡتَدُونَ ٣ ٱللَّهُ ٱلَّذِي خَلَقَ ٱلسَّمَٰوَٰتِ وَٱلۡأَرۡضَ وَمَا بَيۡنَهُمَا فِي سِتَّةِ أَيَّامٖ ثُمَّ ٱسۡتَوَىٰ عَلَى ٱلۡعَرۡشِۖ مَا لَكُم مِّن دُونِهِۦ مِن وَلِيّٖ وَلَا شَفِيعٍۚ أَفَلَا تَتَذَكَّرُونَ ٤ يُدَبِّرُ ٱلۡأَمۡرَ مِنَ ٱلۡسَّمَآءِ إِلَى ٱلۡأَرۡضِ ثُمَّ يَعۡرُجُ إِلَيۡهِ فِي يَوۡمٖ كَانَ مِقۡدَارُهُۥٓ أَلۡفَ سَنَةٖ مِّمَّا تَعُدُّونَ ٥ ذَٰلِكَ عَٰلِمُ ٱلۡغَيۡبِ وَٱلشَّهَٰدَةِ ٱلۡعَزِيزُ ٱلۡرَّحِيمُ ٦ ٱلَّذِيٓ أَحۡسَنَ كُلَّ شَيۡءٍ خَلَقَهُۥۖ وَبَدَأَ خَلۡقَ ٱلۡإِنسَٰنِ مِن طِينٖ ٧ ثُمَّ جَعَلَ نَسۡلَهُۥ مِن سُلَٰلَةٖ مِّن مَّآءٖ مَّهِينٖ ٨ ثُمَّ سَوَّىٰهُ وَنَفَخَ فِيهِ مِن رُّوحِهِۦۖ وَجَعَلَ لَكُمُ ٱلۡسَّمۡعَ وَٱلۡأَبۡصَٰرَ وَٱلۡأَفۡـِٔدَةَۚ قَلِيلٗا مَّا تَشۡكُرُونَ ٩ وَقَالُوٓاْ أَءِذَا ضَلَلۡنَا فِي ٱلۡأَرۡضِ أَءِنَّا لَفِي خَلۡقٖ جَدِيدِۭۚ بَلۡ هُم بِلِقَآءِ رَبِّهِمۡ كَٰفِرُونَ ١٠ ۞قُلۡ يَتَوَفَّىٰكُم مَّلَكُ ٱلۡمَوۡتِ ٱلَّذِي وُكِّلَ بِكُمۡ ثُمَّ إِلَىٰ رَبِّكُمۡ تُرۡجَعُونَ ١١ وَلَوۡ تَرَىٰٓ إِذِ ٱلۡمُجۡرِمُونَ نَاكِسُواْ رُءُوسِهِمۡ عِندَ رَبِّهِمۡ رَبَّنَآ أَبۡصَرۡنَا وَسَمِعۡنَا فَٱرۡجِعۡنَا نَعۡمَلۡ صَٰلِحًا إِنَّا مُوقِنُونَ ١٢ وَلَوۡ شِئۡنَا لَأٓتَيۡنَا كُلَّ نَفۡسٍ هُدَىٰهَا وَلَٰكِنۡ حَقَّ ٱلۡقَوۡلُ مِنِّي لَأَمۡلَأَنَّ جَهَنَّمَ مِنَ ٱلۡجِنَّةِ وَٱلنَّاسِ أَجۡمَعِينَ ١٣ فَذُوقُواْ بِمَا نَسِيتُمۡ لِقَآءَ يَوۡمِكُمۡ هَٰذَآ إِنَّا نَسِينَٰكُمۡۖ وَذُوقُواْ عَذَابَ ٱلۡخُلۡدِ بِمَا كُنتُمۡ تَعۡمَلُونَ ١٤ إِنَّمَا يُؤۡمِنُ بِـَٔايَٰتِنَا ٱلَّذِينَ إِذَا ذُكِّرُواْ بِهَا خَرُّواْۤ سُجَّدٗاۤ وَسَبَّحُواْ بِحَمۡدِ رَبِّهِمۡ وَهُمۡ لَا يَسۡتَكۡبِرُونَ۩ ١٥ تَتَجَافَىٰ جُنُوبُهُمۡ عَنِ ٱلۡمَضَاجِعِ يَدۡعُونَ رَبَّهُمۡ خَوۡفٗا وَطَمَعٗا وَمِمَّا رَزَقۡنَٰهُمۡ يُنفِقُونَ ١٦ فَلَا تَعۡلَمُ نَفۡسٞ مَّآ أُخۡفِيَ لَهُم مِّن قُرَّةِ أَعۡيُنٖ جَزَآءَۢ بِمَا كَانُواْ يَعۡمَلُونَ ١٧ أَفَمَن كَانَ مُؤۡمِنٗا كَمَن كَانَ فَاسِقٗاۚ لَّا يَسۡتَوُۥنَ ١٨ أَمَّا ٱلَّذِينَ ءَامَنُواْ وَعَمِلُواْ ٱلۡصَّٰلِحَٰتِ فَلَهُمۡ جَنَّٰتُ ٱلۡمَأۡوَىٰ نُزُلَۢا بِمَا كَانُواْ يَعۡمَلُونَ ١٩ وَأَمَّا ٱلَّذِينَ فَسَقُواْ فَمَأۡوَىٰهُمُ ٱلۡنَّارُۖ كُلَّمَآ أَرَادُوٓاْ أَن يَخۡرُجُواْ مِنۡهَآ أُعِيدُواْ فِيهَا وَقِيلَ لَهُمۡ ذُوقُواْ عَذَابَ ٱلۡنَّارِ ٱلَّذِي كُنتُم بِهِۦ تُكَذِّبُونَ ٢٠ وَلَنُذِيقَنَّهُم مِّنَ ٱلۡعَذَابِ ٱلۡأَدۡنَىٰ دُونَ ٱلۡعَذَابِ ٱلۡأَكۡبَرِ لَعَلَّهُمۡ يَرۡجِعُونَ ٢١ وَمَنۡ أَظۡلَمُ مِمَّن ذُكِّرَ بِـَٔايَٰتِ رَبِّهِۦ ثُمَّ أَعۡرَضَ عَنۡهَآۚ إِنَّا مِنَ ٱلۡمُجۡرِمِينَ مُنتَقِمُونَ ٢٢ وَلَقَدۡ ءَاتَيۡنَا مُوسَى ٱلۡكِتَٰبَ فَلَا تَكُن فِي مِرۡيَةٖ مِّن لِّقَآئِهِۦۖ وَجَعَلۡنَٰهُ هُدٗى لِّبَنِيٓ إِسۡرَٰٓءِيلَ ٢٣ وَجَعَلۡنَا مِنۡهُمۡ أَئِمَّةٗ يَهۡدُونَ بِأَمۡرِنَا لَمَّا صَبَرُواْۖ وَكَانُواْ بِـَٔايَٰتِنَا يُوقِنُونَ ٢٤ إِنَّ رَبَّكَ هُوَ يَفۡصِلُ بَيۡنَهُمۡ يَوۡمَ ٱلۡقِيَٰمَةِ فِيمَا كَانُواْ فِيهِ يَخۡتَلِفُونَ ٢٥ أَوَلَمۡ يَهۡدِ لَهُمۡ كَمۡ أَهۡلَكۡنَا مِن قَبۡلِهِم مِّنَ ٱلۡقُرُونِ يَمۡشُونَ فِي مَسَٰكِنِهِمۡۚ إِنَّ فِي ذَٰلِكَ لَأٓيَٰتٍۚ أَفَلَا يَسۡمَعُونَ ٢٦ أَوَلَمۡ يَرَوۡاْ أَنَّا نَسُوقُ ٱلۡمَآءَ إِلَى ٱلۡأَرۡضِ ٱلۡجُرُزِ فَنُخۡرِجُ بِهِۦ زَرۡعٗا تَأۡكُلُ مِنۡهُ أَنۡعَٰمُهُمۡ وَأَنفُسُهُمۡۚ أَفَلَا يُبۡصِرُونَ ٢٧ وَيَقُولُونَ مَتَىٰ هَٰذَا ٱلۡفَتۡحُ إِن كُنتُمۡ صَٰدِقِينَ ٢٨ قُلۡ يَوۡمَ ٱلۡفَتۡحِ لَا يَنفَعُ ٱلَّذِينَ كَفَرُوٓاْ إِيمَٰنُهُمۡ وَلَا هُمۡ يُنظَرُونَ ٢٩ فَأَعۡرِضۡ عَنۡهُمۡ وَٱنتَظِرۡ إِنَّهُم مُّنتَظِرُونَ ٣٠''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''بسم الله الرحمن الرحيم
+تَبَٰرَكَ ٱلَّذِي بِيَدِهِ ٱلۡمُلۡكُ وَهُوَ عَلَىٰ كُلِّ شَيۡءٖ قَدِيرٌ ١ ٱلَّذِي خَلَقَ ٱلۡمَوۡتَ وَٱلۡحَيَوٰةَ لِيَبۡلُوَكُمۡ أَيُّكُمۡ أَحۡسَنُ عَمَلٗاۚ وَهُوَ ٱلۡعَزِيزُ ٱلۡغَفُورُ ٢ ٱلَّذِي خَلَقَ سَبۡعَ سَمَٰوَٰتٖ طِبَاقٗاۖ مَّا تَرَىٰ فِي خَلۡقِ ٱلرَّحۡمَٰنِ مِن تَفَٰوُتٖۖ فَٱرۡجِعِ ٱلۡبَصَرَ هَلۡ تَرَىٰ مِن فُطُورٖ ٣ ثُمَّ ٱرۡجِعِ ٱلۡبَصَرَ كَرَّتَيۡنِ يَنقَلِبۡ إِلَيۡكَ ٱلۡبَصَرُ خَاسِئٗا وَهُوَ حَسِيرٞ ٤ وَلَقَدۡ زَيَّنَّا ٱلۡسَّمَآءَ ٱلدُّنۡيَا بِمَصَٰبِيحَ وَجَعَلۡنَٰهَا رُجُومٗا لِّلشَّيَٰطِينِۖ وَأَعۡتَدۡنَا لَهُمۡ عَذَابَ ٱلۡسَّعِيرِ ٥ وَلِلَّذِينَ كَفَرُواْ بِرَبِّهِمۡ عَذَابُ جَهَنَّمَۖ وَبِئۡسَ ٱلۡمَصِيرُ ٦ إِذَآ أُلۡقُواْ فِيهَا سَمِعُواْ لَهَا شَهِيقٗا وَهِيَ تَفُورُ ٧ تَكَادُ تَمَيَّزُ مِنَ ٱلۡغَيۡظِۖ كُلَّمَآ أُلۡقِيَ فِيهَا فَوۡجٞ سَأَلَهُمۡ خَزَنَتُهَآ أَلَمۡ يَأۡتِكُمۡ نَذِيرٞ ٨ قَالُواْ بَلَىٰ قَدۡ جَآءَنَا نَذِيرٞ فَكَذَّبۡنَا وَقُلۡنَا مَا نَزَّلَ ٱللَّهُ مِن شَيۡءٍ إِنۡ أَنتُمۡ إِلَّا فِي ضَلَٰلٖ كَبِيرٖ ٩ وَقَالُواْ لَوۡ كُنَّا نَسۡمَعُ أَوۡ نَعۡقِلُ مَا كُنَّا فِيٓ أَصۡحَٰبِ ٱلۡسَّعِيرِ ١٠ فَٱعۡتَرَفُواْ بِذَنۢبِهِمۡ فَسُحۡقٗا لِّأَصۡحَٰبِ ٱلۡسَّعِيرِ ١١ إِنَّ ٱلَّذِينَ يَخۡشَوۡنَ رَبَّهُم بِٱلۡغَيۡبِ لَهُم مَّغۡفِرَةٞ وَأَجۡرٞ كَبِيرٞ ١٢ وَأَسِرُّواْ قَوۡلَكُمۡ أَوِ ٱجۡهَرُواْ بِهِۦٓۖ إِنَّهُۥ عَلِيمُۢ بِذَاتِ ٱلصُّدُورِ ١٣ أَلَا يَعۡلَمُ مَنۡ خَلَقَ وَهُوَ ٱللَّطِيفُ ٱلۡخَبِيرُ ١٤ هُوَ ٱلَّذِي جَعَلَ لَكُمُ ٱلۡأَرۡضَ ذَلُولٗا فَٱمۡشُواْ فِي مَنَاكِبِهَا وَكُلُواْ مِن رِّزۡقِهِۦۖ وَإِلَيۡهِ ٱلنُّشُورُ ١٥ ءَأَمِنتُم مَّن فِي ٱلۡسَّمَآءِ أَن يَخۡسِفَ بِكُمُ ٱلۡأَرۡضَ فَإِذَا هِيَ تَمُورُ ١٦ أَمۡ أَمِنتُم مَّن فِي ٱلۡسَّمَآءِ أَن يُرۡسِلَ عَلَيۡكُمۡ حَاصِبٗاۖ فَسَتَعۡلَمُونَ كَيۡفَ نَذِيرِ ١٧ وَلَقَدۡ كَذَّبَ ٱلَّذِينَ مِن قَبۡلِهِمۡ فَكَيۡفَ كَانَ نَكِيرِ ١٨ أَوَلَمۡ يَرَوۡاْ إِلَى ٱلطَّيۡرِ فَوۡقَهُمۡ صَٰٓفَّٰتٖ وَيَقۡبِضۡنَۚ مَا يُمۡسِكُهُنَّ إِلَّا ٱلرَّحۡمَٰنُۚ إِنَّهُۥ بِكُلِّ شَيۡءِۭ بَصِيرٌ ١٩ أَمَّنۡ هَٰذَا ٱلَّذِي هُوَ جُندٞ لَّكُمۡ يَنصُرُكُم مِّن دُونِ ٱلرَّحۡمَٰنِۚ إِنِ ٱلۡكَٰفِرُونَ إِلَّا فِي غُرُورٍ ٢٠ أَمَّنۡ هَٰذَا ٱلَّذِي يَرۡزُقُكُمۡ إِنۡ أَمۡسَكَ رِزۡقَهُۥۚ بَل لَّجُّواْ فِي عُتُوّٖ وَنُفُورٍ ٢١ أَفَمَن يَمۡشِي مُكِبًّا عَلَىٰ وَجۡهِهِۦٓ أَهۡدَىٰٓ أَمَّن يَمۡشِي سَوِيًّا عَلَىٰ صِرَٰطٖ مُّسۡتَقِيمٖ ٢٢ قُلۡ هُوَ ٱلَّذِيٓ أَنشَأَكُمۡ وَجَعَلَ لَكُمُ ٱلۡسَّمۡعَ وَٱلۡأَبۡصَٰرَ وَٱلۡأَفۡـِٔدَةَۚ قَلِيلٗا مَّا تَشۡكُرُونَ ٢٣ قُلۡ هُوَ ٱلَّذِي ذَرَأَكُمۡ فِي ٱلۡأَرۡضِ وَإِلَيۡهِ تُحۡشَرُونَ ٢٤ وَيَقُولُونَ مَتَىٰ هَٰذَا ٱلۡوَعۡدُ إِن كُنتُمۡ صَٰدِقِينَ ٢٥ قُلۡ إِنَّمَا ٱلۡعِلۡمُ عِندَ ٱللَّهِ وَإِنَّمَآ أَنَا۠ نَذِيرٞ مُّبِينٞ ٢٦ فَلَمَّا رَأَوۡهُ زُلۡفَةٗ سِيٓـَٔتۡ وُجُوهُ ٱلَّذِينَ كَفَرُواْ وَقِيلَ هَٰذَا ٱلَّذِي كُنتُم بِهِۦ تَدَّعُونَ ٢٧ قُلۡ أَرَءَيۡتُمۡ إِنۡ أَهۡلَكَنِيَ ٱللَّهُ وَمَن مَّعِيَ أَوۡ رَحِمَنَا فَمَن يُجِيرُ ٱلۡكَٰفِرِينَ مِنۡ عَذَابٍ أَلِيمٖ ٢٨ قُلۡ هُوَ ٱلرَّحۡمَٰنُ ءَامَنَّا بِهِۦ وَعَلَيۡهِ تَوَكَّلۡنَاۖ فَسَتَعۡلَمُونَ مَنۡ هُوَ فِي ضَلَٰلٖ مُّبِينٖ ٢٩ قُلۡ أَرَءَيۡتُمۡ إِنۡ أَصۡبَحَ مَآؤُكُمۡ غَوۡرٗا فَمَن يَأۡتِيكُم بِمَآءٖ مَّعِينِۭ ٣٠''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''اللَّهُمَّ أَسْلَمْتُ نَفْسِي إِلَيْكَ، وَفَوَّضْتُ أَمْرِي إِلَيْكَ، وَوَجَّهْتُ وَجْهِي إِلَيْكَ، وَأَلْجَأْتُ ظَهْرِي إِلَيْكَ، رَغْبَةً وَرَهْبَةً إِلَيْكَ، لاَ مَلْجَأَ وَلاَ مَنْجَا مِنْكَ إِلاَّ إِلَيْكَ، آمَنْتُ بِكِتَابِكَ الَّذِي أَنْزَلْتَ، وَبِنَبِيِّكَ الَّذِي أَرْسَلْتَ''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+];
+
+///
+/// ──────────────────────────────────────────────────────────────────────────────
+/// WAKING UP AZKAR
+/// ──────────────────────────────────────────────────────────────────────────────
+final List<DhikrItem> wakingUpAzkar = [
+  DhikrItem(
+    arabic: '''الْحَمْدُ للَّهِ الَّذِي أَحْيَانَا بَعْدَ مَا أَمَاتَنَا، وَإِلَيْهِ النُّشُورُ''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''لاَ إِلَهَ إِلاَّ اللَّهُ وَحْدَهُ لاَ شَريكَ لَهُ، لَهُ الْمُلْكُ وَلَهُ الْحَمْدُ، وَهُوَ عَلَى كُلِّ شَيْءٍ قَدِيرٌ، سُبْحَانَ اللَّهِ، وَالْحَمْدُ للَّهِ، وَلاَ إِلَهَ إِلاَّ اللَّهُ، وَاللَّهُ أَكبَرُ، وَلاَ حَوْلَ وَلاَ قُوَّةَ إِلاَّ بِاللَّهِ الْعَلِيِّ الْعَظِيمِ، رَبِّ اغْفرْ لِي''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''الْحَمْدُ لِلَّهِ الَّذِي عَافَانِي فِي جَسَدِي، وَرَدَّ عَلَيَّ رُوحِي، وَأَذِنَ لي بِذِكْرِهِ''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''إِنَّ فِي خَلۡقِ ٱلسَّمَٰوَٰتِ وَٱلۡأَرۡضِ وَٱخۡتِلَٰفِ ٱلَّيۡلِ وَٱلنَّهَارِ لَأٓيَٰتٖ لِّأُوْلِي ٱلۡأَلۡبَٰبِ ١٩٠ ٱلَّذِينَ يَذۡكُرُونَ ٱللَّهَ قِيَٰمٗا وَقُعُودٗا وَعَلَىٰ جُنُوبِهِمۡ وَيَتَفَكَّرُونَ فِي خَلۡقِ ٱلسَّمَٰوَٰتِ وَٱلۡأَرۡضِ رَبَّنَا مَا خَلَقۡتَ هَٰذَا بَٰطِلٗا سُبۡحَٰنَكَ فَقِنَا عَذَابَ ٱلنَّارِ ١٩١ رَبَّنَآ إِنَّكَ مَن تُدۡخِلِ ٱلنَّارَ فَقَدۡ أَخۡزَيۡتَهُۥۖ وَمَا لِلظَّٰلِمِينَ مِنۡ أَنصَارٖ ١٩٢ رَّبَّنَآ إِنَّنَا سَمِعۡنَا مُنَادِيٗا يُنَادِي لِلۡإِيمَٰنِ أَنۡ ءَامِنُواْ بِرَبِّكُمۡ فَـَٔامَنَّاۚ رَبَّنَا فَٱغۡفِرۡ لَنَا ذُنُوبَنَا وَكَفِّرۡ عَنَّا سَيِّـَٔاتِنَا وَتَوَفَّنَا مَعَ ٱلۡأَبۡرَارِ ١٩٣ رَبَّنَا وَءَاتِنَا مَا وَعَدتَّنَا عَلَىٰ رُسُلِكَ وَلَا تُخۡزِنَا يَوۡمَ ٱلۡقِيَٰمَةِۖ إِنَّكَ لَا تُخۡلِفُ ٱلۡمِيعَادَ ١٩٤ فَٱسۡتَجَابَ لَهُمۡ رَبُّهُمۡ أَنِّي لَآ أُضِيعُ عَمَلَ عَٰمِلٖ مِّنكُم مِّن ذَكَرٍ أَوۡ أُنثَىٰۖ بَعۡضُكُم مِّنۢ بَعۡضٖۖ فَٱلَّذِينَ هَاجَرُواْ وَأُخۡرِجُواْ مِن دِيَٰرِهِمۡ وَأُوذُواْ فِي سَبِيلِي وَقَٰتَلُواْ وَقُتِلُواْ لَأُكَفِّرَنَّ عَنۡهُمۡ سَيِّـَٔاتِهِمۡ وَلَأُدۡخِلَنَّهُمۡ جَنَّٰتٖ تَجۡرِي مِن تَحۡتِهَا ٱلۡأَنۡهَٰرُ ثَوَابٗا مِّنۡ عِندِ ٱللَّهِۚ وَٱللَّهُ عِندَهُۥ حُسۡنُ ٱلثَّوَابِ ١٩٥ لَا يَغُرَّنَّكَ تَقَلُّبُ ٱلَّذِينَ كَفَرُواْ فِي ٱلۡبِلَٰدِ ١٩٦ مَتَٰعٞ قَلِيلٞ ثُمَّ مَأۡوَىٰهُمۡ جَهَنَّمُۖ وَبِئۡسَ ٱلۡمِهَادُ ١٩٧ لَٰكِنِ ٱلَّذِينَ ٱتَّقَوۡاْ رَبَّهُمۡ لَهُمۡ جَنَّٰتٞ تَجۡرِي مِن تَحۡتِهَا ٱلۡأَنۡهَٰرُ خَٰلِدِينَ فِيهَا نُزُلٗا مِّنۡ عِندِ ٱللَّهِۗ وَمَا عِندَ ٱللَّهِ خَيۡرٞ لِّلۡأَبۡرَارِ ١٩٨ وَإِنَّ مِنۡ أَهۡلِ ٱلۡكِتَٰبِ لَمَن يُؤۡمِنُ بِٱللَّهِ وَمَآ أُنزِلَ إِلَيۡكُمۡ وَمَآ أُنزِلَ إِلَيۡهِمۡ خَٰشِعِينَ لِلَّهِ لَا يَشۡتَرُونَ بِـَٔايَٰتِ ٱللَّهِ ثَمَنٗا قَلِيلًاۚ أُوْلَٰٓئِكَ لَهُمۡ أَجۡرُهُمۡ عِندَ رَبِّهِمۡۗ إِنَّ ٱللَّهَ سَرِيعُ ٱلۡحِسَابِ ١٩٩ يَٰٓأَيُّهَا ٱلَّذِينَ ءَامَنُواْ ٱصۡبِرُواْ وَصَابِرُواْ وَرَابِطُواْ وَٱتَّقُواْ ٱللَّهَ لَعَلَّكُمۡ تُفۡلِحُونَ ٢٠٠''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+];
+
+///
+/// ──────────────────────────────────────────────────────────────────────────────
+/// AFTER PRAYERS
+/// ──────────────────────────────────────────────────────────────────────────────
+final List<DhikrItem> afterPrayersAzkar = [
+  DhikrItem(
+    arabic: '''أَسْتَغْفِرُ اللَّهَ , أَسْتَغْفِرُ اللَّهَ  , أَسْتَغْفِرُ اللَّهَ.
+
+اللَّهُمَّ أَنْتَ السَّلاَمُ، وَمِنْكَ السَّلاَمُ، تَبَارَكْتَ يَا ذَا الْجَلاَلِ وَالْإِكْرَامِ''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''لاَ إِلَهَ إِلاَّ اللَّهُ وَحْدَهُ لاَ شَرِيكَ لَهُ، لَهُ الْمُلْكُ وَلَهُ الْحَمْدُ وَهُوَ عَلَى كُلِّ شَيْءٍ قَدِيرٌ، اللَّهُمَّ لاَ مَانِعَ لِمَا أَعْطَيْتَ، وَلاَ مُعْطِيَ لِمَا مَنَعْتَ، وَلاَ يَنْفَعُ ذَا الْجَدِّ مِنْكَ الجَدُّ''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''لَا إِلَهَ إِلاَّ اللَّهُ وَحْدَهُ لاَ شَرِيكَ لَهُ، لَهُ الْمُلْكُ، وَلَهُ الْحَمدُ، وَهُوَ عَلَى كُلِّ شَيْءٍ قَدِيرٌ. لاَ حَوْلَ وَلاَ قُوَّةَ إِلاَّ بِاللَّهِ، لاَ إِلَهَ إِلاَّ اللَّهُ، وَلاَ نَعْبُدُ إِلاَّ إِيَّاهُ, لَهُ النِّعْمَةُ وَلَهُ الْفَضْلُ وَلَهُ الثَّنَاءُ الْحَسَنُ، لَا إِلَهَ إِلاَّ اللَّهُ مُخْلِصِينَ لَهُ الدِّينَ وَلَوْ كَرِهَ الكَافِرُونَ''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''سُـبْحانَ اللهِ، والحَمْـدُ لله ، واللهُ أكْـبَر.''',
+    translation: '33 time',
+    repeat: 33,
+  ),
+  DhikrItem(
+    arabic: '''لاَ إِلَهَ إِلاَّ اللَّهُ وَحْدَهُ لاَ شَرِيكَ لَهُ، لَهُ الْمُلْكُ وَلَهُ الْحَمْدُ وَهُوَ عَلَى كُلِّ شَيْءٍ قَدِيرٌ''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''بسم الله الرحمن الرحيم
+قُلۡ هُوَ ٱللَّهُ أَحَدٌ ١ ٱللَّهُ ٱلصَّمَدُ ٢ لَمۡ يَلِدۡ وَلَمۡ يُولَدۡ ٣ وَلَمۡ يَكُن لَّهُۥ كُفُوًا أَحَدُۢ ٤''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''بسم الله الرحمن الرحيم
+قُلۡ أَعُوذُ بِرَبِّ ٱلۡفَلَقِ ١ مِن شَرِّ مَا خَلَقَ ٢ وَمِن شَرِّ غَاسِقٍ إِذَا وَقَبَ ٣ وَمِن شَرِّ ٱلنَّفَّٰثَٰتِ فِي ٱلۡعُقَدِ ٤ وَمِن شَرِّ حَاسِدٍ إِذَا حَسَدَ ٥''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''بسم الله الرحمن الرحيم
+قُلۡ أَعُوذُ بِرَبِّ ٱلنَّاسِ ١ مَلِكِ ٱلنَّاسِ ٢ إِلَٰهِ ٱلنَّاسِ ٣ مِن شَرِّ ٱلۡوَسۡوَاسِ ٱلۡخَنَّاسِ ٤ ٱلَّذِي يُوَسۡوِسُ فِي صُدُورِ ٱلنَّاسِ ٥ مِنَ ٱلۡجِنَّةِ وَٱلنَّاسِ ٦''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''ٱللَّهُ لَآ إِلَٰهَ إِلَّا هُوَ ٱلۡحَيُّ ٱلۡقَيُّومُۚ لَا تَأۡخُذُهُۥ سِنَةٞ وَلَا نَوۡمٞۚ لَّهُۥ مَا فِي ٱلسَّمَٰوَٰتِ وَمَا فِي ٱلۡأَرۡضِۗ مَن ذَا ٱلَّذِي يَشۡفَعُ عِندَهُۥٓ إِلَّا بِإِذۡنِهِۦۚ يَعۡلَمُ مَا بَيۡنَ أَيۡدِيهِمۡ وَمَا خَلۡفَهُمۡۖ وَلَا يُحِيطُونَ بِشَيۡءٖ مِّنۡ عِلۡمِهِۦٓ إِلَّا بِمَا شَآءَۚ وَسِعَ كُرۡسِيُّهُ ٱلۡسَّمَٰوَٰتِ وَٱلۡأَرۡضَۖ وَلَا يَـُٔودُهُۥ حِفۡظُهُمَاۚ وَهُوَ ٱلۡعَلِيُّ ٱلۡعَظِيمُ ٢٥٥''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: '''لاَ إِلَهَ إِلاَّ اللَّهُ وَحْدَهُ لاَ شَرِيكَ لَهُ، لَهُ الْمُلْكُ وَلَهُ الْحَمْدُ يُحْيِي وَيُمِيتُ وَهُوَ عَلَى كُلِّ شَيْءٍ قَدِيرٌ''',
+    translation: '10 times',
+    repeat: 10,
+  ),
+  DhikrItem(
+    arabic: '''اللَّهُمَّ إِنِّي أَسْأَلُكَ عِلْماً نافِعاً، وَرِزْقاً طَيِّباً، وَعَمَلاً مُتَقَبَّلاً''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+];
+
+///
+/// ──────────────────────────────────────────────────────────────────────────────
+/// SURAH AL MULK (Separate Listing As Provided At The End)
+/// ──────────────────────────────────────────────────────────────────────────────
+/// You gave the same Surah Al-Mulk text again at the end. We store it separately.
+final List<DhikrItem> surahAlMulkAzkar = [
+  DhikrItem(
+    arabic: '''بسم الله الرحمن الرحيم
+تَبَٰرَكَ ٱلَّذِي بِيَدِهِ ٱلۡمُلۡكُ وَهُوَ عَلَىٰ كُلِّ شَيۡءٖ قَدِيرٌ ١ ٱلَّذِي خَلَقَ ٱلۡمَوۡتَ وَٱلۡحَيَوٰةَ لِيَبۡلُوَكُمۡ أَيُّكُمۡ أَحۡسَنُ عَمَلٗاۚ وَهُوَ ٱلۡعَزِيزُ ٱلۡغَفُورُ ٢ ٱلَّذِي خَلَقَ سَبۡعَ سَمَٰوَٰتٖ طِبَاقٗاۖ مَّا تَرَىٰ فِي خَلۡقِ ٱلرَّحۡمَٰنِ مِن تَفَٰوُتٖۖ فَٱرۡجِعِ ٱلۡبَصَرَ هَلۡ تَرَىٰ مِن فُطُورٖ ٣ ثُمَّ ٱرۡجِعِ ٱلۡبَصَرَ كَرَّتَيۡنِ يَنقَلِبۡ إِلَيۡكَ ٱلۡبَصَرُ خَاسِئٗا وَهُوَ حَسِيرٞ ٤ وَلَقَدۡ زَيَّنَّا ٱلۡسَّمَآءَ ٱلدُّنۡيَا بِمَصَٰبِيحَ وَجَعَلۡنَٰهَا رُجُومٗا لِّلشَّيَٰطِينِۖ وَأَعۡتَدۡنَا لَهُمۡ عَذَابَ ٱلۡسَّعِيرِ ٥ وَلِلَّذِينَ كَفَرُواْ بِرَبِّهِمۡ عَذَابُ جَهَنَّمَۖ وَبِئۡسَ ٱلۡمَصِيرُ ٦ إِذَآ أُلۡقُواْ فِيهَا سَمِعُواْ لَهَا شَهِيقٗا وَهِيَ تَفُورُ ٧ تَكَادُ تَمَيَّزُ مِنَ ٱلۡغَيۡظِۖ كُلَّمَآ أُلۡقِيَ فِيهَا فَوۡجٞ سَأَلَهُمۡ خَزَنَتُهَآ أَلَمۡ يَأۡتِكُمۡ نَذِيرٞ ٨ قَالُواْ بَلَىٰ قَدۡ جَآءَنَا نَذِيرٞ فَكَذَّبۡنَا وَقُلۡنَا مَا نَزَّلَ ٱللَّهُ مِن شَيۡءٍ إِنۡ أَنتُمۡ إِلَّا فِي ضَلَٰلٖ كَبِيرٖ ٩ وَقَالُواْ لَوۡ كُنَّا نَسۡمَعُ أَوۡ نَعۡقِلُ مَا كُنَّا فِيٓ أَصۡحَٰبِ ٱلۡسَّعِيرِ ١٠ فَٱعۡتَرَفُواْ بِذَنۢبِهِمۡ فَسُحۡقٗا لِّأَصۡحَٰبِ ٱلۡسَّعِيرِ ١١ إِنَّ ٱلَّذِينَ يَخۡشَوۡنَ رَبَّهُم بِٱلۡغَيۡبِ لَهُم مَّغۡفِرَةٞ وَأَجۡرٞ كَبِيرٞ ١٢ وَأَسِرُّواْ قَوۡلَكُمۡ أَوِ ٱجۡهَرُواْ بِهِۦٓۖ إِنَّهُۥ عَلِيمُۢ بِذَاتِ ٱلصُّدُورِ ١٣ أَلَا يَعۡلَمُ مَنۡ خَلَقَ وَهُوَ ٱللَّطِيفُ ٱلۡخَبِيرُ ١٤ هُوَ ٱلَّذِي جَعَلَ لَكُمُ ٱلۡأَرۡضَ ذَلُولٗا فَٱمۡشُواْ فِي مَنَاكِبِهَا وَكُلُواْ مِن رِّزۡقِهِۦۖ وَإِلَيۡهِ ٱلنُّشُورُ ١٥ ءَأَمِنتُم مَّن فِي ٱلۡسَّمَآءِ أَن يَخۡسِفَ بِكُمُ ٱلۡأَرۡضَ فَإِذَا هِيَ تَمُورُ ١٦ أَمۡ أَمِنتُم مَّن فِي ٱلۡسَّمَآءِ أَن يُرۡسِلَ عَلَيۡكُمۡ حَاصِبٗاۖ فَسَتَعۡلَمُونَ كَيۡفَ نَذِيرِ ١٧ وَلَقَدۡ كَذَّبَ ٱلَّذِينَ مِن قَبۡلِهِمۡ فَكَيۡفَ كَانَ نَكِيرِ ١٨ أَوَلَمۡ يَرَوۡاْ إِلَى ٱلطَّيۡرِ فَوۡقَهُمۡ صَٰٓفَّٰتٖ وَيَقۡبِضۡنَۚ مَا يُمۡسِكُهُنَّ إِلَّا ٱلرَّحۡمَٰنُۚ إِنَّهُۥ بِكُلِّ شَيۡءِۭ بَصِيرٌ ١٩ أَمَّنۡ هَٰذَا ٱلَّذِي هُوَ جُندٞ لَّكُمۡ يَنصُرُكُم مِّن دُونِ ٱلرَّحۡمَٰنِۚ إِنِ ٱلۡكَٰفِرُونَ إِلَّا فِي غُرُورٍ ٢٠ أَمَّنۡ هَٰذَا ٱلَّذِي يَرۡزُقُكُمۡ إِنۡ أَمۡسَكَ رِزۡقَهُۥۚ بَل لَّجُّواْ فِي عُتُوّٖ وَنُفُورٍ ٢١ أَفَمَن يَمۡشِي مُكِبًّا عَلَىٰ وَجۡهِهِۦٓ أَهۡدَىٰٓ أَمَّن يَمۡشِي سَوِيًّا عَلَىٰ صِرَٰطٖ مُّسۡتَقِيمٖ ٢٢ قُلۡ هُوَ ٱلَّذِيٓ أَنشَأَكُمۡ وَجَعَلَ لَكُمُ ٱلۡسَّمۡعَ وَٱلۡأَبۡصَٰرَ وَٱلۡأَفۡـِٔدَةَۚ قَلِيلٗا مَّا تَشۡكُرُونَ ٢٣ قُلۡ هُوَ ٱلَّذِي ذَرَأَكُمۡ فِي ٱلۡأَرۡضِ وَإِلَيۡهِ تُحۡشَرُونَ ٢٤ وَيَقُولُونَ مَتَىٰ هَٰذَا ٱلۡوَعۡدُ إِن كُنتُمۡ صَٰدِقِينَ ٢٥ قُلۡ إِنَّمَا ٱلۡعِلۡمُ عِندَ ٱللَّهِ وَإِنَّمَآ أَنَا۠ نَذِيرٞ مُّبِينٞ ٢٦ فَلَمَّا رَأَوۡهُ زُلۡفَةٗ سِيٓـَٔتۡ وُجُوهُ ٱلَّذِينَ كَفَرُواْ وَقِيلَ هَٰذَا ٱلَّذِي كُنتُم بِهِۦ تَدَّعُونَ ٢٧ قُلۡ أَرَءَيۡتُمۡ إِنۡ أَهۡلَكَنِيَ ٱللَّهُ وَمَن مَّعِيَ أَوۡ رَحِمَنَا فَمَن يُجِيرُ ٱلۡكَٰفِرِينَ مِنۡ عَذَابٍ أَلِيمٖ ٢٨ قُلۡ هُوَ ٱلرَّحۡمَٰنُ ءَامَنَّا بِهِۦ وَعَلَيۡهِ تَوَكَّلۡنَاۖ فَسَتَعۡلَمُونَ مَنۡ هُوَ فِي ضَلَٰلٖ مُّبِينٖ ٢٩ قُلۡ أَرَءَيۡتُمۡ إِنۡ أَصۡبَحَ مَآؤُكُمۡ غَوۡرٗا فَمَن يَأۡتِيكُم بِمَآءٖ مَّعِينِۭ ٣٠''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+];
+
+///
+/// ──────────────────────────────────────────────────────────────────────────────
+/// SURAH YASEEN (Separate Listing)
+/// ──────────────────────────────────────────────────────────────────────────────
+final List<DhikrItem> surahYaseenAzkar = [
+  DhikrItem(
+    arabic: '''بسم الله الرحمن الرحيم
+يسٓ ١ وَٱلۡقُرۡءَانِ ٱلۡحَكِيمِ ٢ إِنَّكَ لَمِنَ ٱلۡمُرۡسَلِينَ ٣ عَلَىٰ صِرَٰطٖ مُّسۡتَقِيمٖ ٤ تَنزِيلَ ٱلۡعَزِيزِ ٱلرَّحِيمِ ٥ لِتُنذِرَ قَوۡمٗا مَّآ أُنذِرَ ءَابَآؤُهُمۡ فَهُمۡ غَٰفِلُونَ ٦ لَقَدۡ حَقَّ ٱلۡقَوۡلُ عَلَىٰٓ أَكۡثَرِهِمۡ فَهُمۡ لَا يُؤۡمِنُونَ ٧ إِنَّا جَعَلۡنَا فِيٓ أَعۡنَٰقِهِمۡ أَغۡلَٰلٗا فَهِيَ إِلَى ٱلۡأَذۡقَانِ فَهُم مُّقۡمَحُونَ ٨ وَجَعَلۡنَا مِنۢ بَيۡنِ أَيۡدِيهِمۡ سَدّٗا وَمِنۡ خَلۡفِهِمۡ سَدّٗا فَأَغۡشَيۡنَٰهُمۡ فَهُمۡ لَا يُبۡصِرُونَ ٩ وَسَوَآءٌ عَلَيۡهِمۡ ءَأَنذَرۡتَهُمۡ أَمۡ لَمۡ تُنذِرۡهُمۡ لَا يُؤۡمِنُونَ ١٠ إِنَّمَا تُنذِرُ مَنِ ٱتَّبَعَ ٱلذِّكۡرَ وَخَشِيَ ٱلرَّحۡمَٰنَ بِٱلۡغَيۡبِۖ فَبَشِّرۡهُ بِمَغۡفِرَةٖ وَأَجۡرٖ كَرِيمٍ ١١ إِنَّا نَحۡنُ نُحۡيِ ٱلۡمَوۡتَىٰ وَنَكۡتُبُ مَا قَدَّمُواْ وَءَاثَٰرَهُمۡۚ وَكُلَّ شَيۡءٍ أَحۡصَيۡنَٰهُ فِيٓ إِمَامٖ مُّبِينٖ ١٢ وَٱضۡرِبۡ لَهُم مَّثَلًا أَصۡحَٰبَ ٱلۡقَرۡيَةِ إِذۡ جَآءَهَا ٱلۡمُرۡسَلُونَ ١٣ إِذۡ أَرۡسَلۡنَآ إِلَيۡهِمُ ٱثۡنَيۡنِ فَكَذَّبُوهُمَا فَعَزَّزۡنَا بِثَالِثٖ فَقَالُوٓاْ إِنَّآ إِلَيۡكُم مُّرۡسَلُونَ ١٤ قَالُواْ مَآ أَنتُمۡ إِلَّا بَشَرٞ مِّثۡلُنَا وَمَآ أَنزَلَ ٱلرَّحۡمَٰنُ مِن شَيۡءٍ إِنۡ أَنتُمۡ إِلَّا تَكۡذِبُونَ ١٥ قَالُواْ رَبُّنَا يَعۡلَمُ إِنَّآ إِلَيۡكُمۡ لَمُرۡسَلُونَ ١٦ وَمَا عَلَيۡنَآ إِلَّا ٱلۡبَلَٰغُ ٱلۡمُبِينُ ١٧ قَالُوٓاْ إِنَّا تَطَيَّرۡنَا بِكُمۡۖ لَئِن لَّمۡ تَنتَهُواْ لَنَرۡجُمَنَّكُمۡ وَلَيَمَسَّنَّكُم مِّنَّا عَذَابٌ أَلِيمٞ ١٨ قَالُواْ طَٰٓئِرُكُم مَّعَكُمۡ أَئِن ذُكِّرۡتُمۚ بَلۡ أَنتُمۡ قَوۡمٞ مُّسۡرِفُونَ ١٩ وَجَآءَ مِنۡ أَقۡصَا ٱلۡمَدِينَةِ رَجُلٞ يَسۡعَىٰ قَالَ يَٰقَوۡمِ ٱتَّبِعُواْ ٱلۡمُرۡسَلِينَ ٢٠ ٱتَّبِعُواْ مَن لَّا يَسۡـَٔلُكُمۡ أَجۡرٗا وَهُم مُّهۡتَدُونَ ٢١ وَمَالِيَ لَآ أَعۡبُدُ ٱلَّذِي فَطَرَنِي وَإِلَيۡهِ تُرۡجَعُونَ ٢٢ ءَأَتَّخِذُ مِن دُونِهِۦٓ ءَالِهَةً إِن يُرِدۡنِ ٱلرَّحۡمَٰنُ بِضُرّٖ لَّا تُغۡنِ عَنِّي شَفَٰعَتُهُمۡ شَيۡـٔٗا وَلَا يُنقِذُونِ ٢٣ إِنِّيٓ إِذٗا لَّفِي ضَلَٰلٖ مُّبِينٍ ٢٤ إِنِّيٓ ءَامَنتُ بِرَبِّكُمۡ فَٱسۡمَعُونِ ٢٥ قِيلَ ٱدۡخُلِ ٱلۡجَنَّةَۖ قَالَ يَٰلَيۡتَ قَوۡمِي يَعۡلَمُونَ ٢٦ بِمَا غَفَرَ لِي رَبِّي وَجَعَلَنِي مِنَ ٱلۡمُكۡرَمِينَ ٢٧ ۞وَمَآ أَنزَلۡنَا عَلَىٰ قَوۡمِهِۦ مِنۢ بَعۡدِهِۦ مِن جُندٖ مِّنَ ٱلۡسَّمَآءِ وَمَا كُنَّا مُنزِلِينَ ٢٨ إِن كَانَتۡ إِلَّا صَيۡحَةٗ وَٰحِدَةٗ فَإِذَا هُمۡ خَٰمِدُونَ ٢٩ يَٰحَسۡرَةً عَلَى ٱلۡعِبَادِۚ مَا يَأۡتِيهِم مِّن رَّسُولٍ إِلَّا كَانُواْ بِهِۦ يَسۡتَهۡزِءُونَ ٣٠ أَلَمۡ يَرَوۡاْ كَمۡ أَهۡلَكۡنَا قَبۡلَهُم مِّنَ ٱلۡقُرُونِ أَنَّهُمۡ إِلَيۡهِمۡ لَا يَرۡجِعُونَ ٣١ وَإِن كُلّٞ لَّمَّا جَمِيعٞ لَّدَيۡنَا مُحۡضَرُونَ ٣٢ وَءَايَةٞ لَّهُمُ ٱلۡأَرۡضُ ٱلۡمَيۡتَةُ أَحۡيَيۡنَٰهَا وَأَخۡرَجۡنَا مِنۡهَا حَبّٗا فَمِنۡهُ يَأۡكُلُونَ ٣٣ وَجَعَلۡنَا فِيهَا جَنَّٰتٖ مِّن نَّخِيلٖ وَأَعۡنَٰبٖ وَفَجَّرۡنَا فِيهَا مِنَ ٱلۡعُيُونِ ٣٤ لِيَأۡكُلُواْ مِن ثَمَرِهِۦ وَمَا عَمِلَتۡهُ أَيۡدِيهِمۡۚ أَفَلَا يَشۡكُرُونَ ٣٥ سُبۡحَٰنَ ٱلَّذِي خَلَقَ ٱلۡأَزۡوَٰجَ كُلَّهَا مِمَّا تُنۢبِتُ ٱلۡأَرۡضُ وَمِنۡ أَنفُسِهِمۡ وَمِمَّا لَا يَعۡلَمُونَ ٣٦ وَءَايَةٞ لَّهُمُ ٱلَّيۡلُ نَسۡلَخُ مِنۡهُ ٱلنَّهَارَ فَإِذَا هُم مُّظۡلِمُونَ ٣٧ وَٱلشَّمۡسُ تَجۡرِي لِمُسۡتَقَرّٖ لَّهَاۚ ذَٰلِكَ تَقۡدِيرُ ٱلۡعَزِيزِ ٱلۡعَلِيمِ ٣٨ وَٱلۡقَمَرَ قَدَّرۡنَٰهُ مَنَازِلَ حَتَّىٰ عَادَ كَٱلۡعُرۡجُونِ ٱلۡقَدِيمِ ٣٩ لَا ٱلشَّمۡسُ يَنۢبَغِي لَهَآ أَن تُدۡرِكَ ٱلۡقَمَرَ وَلَا ٱلَّيۡلُ سَابِقُ ٱلنَّهَارِۚ وَكُلّٞ فِي فَلَكٖ يَسۡبَحُونَ ٤٠ وَءَايَةٞ لَّهُمۡ أَنَّا حَمَلۡنَا ذُرِّيَّتَهُمۡ فِي ٱلۡفُلۡكِ ٱلۡمَشۡحُونِ ٤١ وَخَلَقۡنَا لَهُم مِّن مِّثۡلِهِۦ مَا يَرۡكَبُونَ ٤٢ وَإِن نَّشَأۡ نُغۡرِقۡهُمۡ فَلَا صَرِيخَ لَهُمۡ وَلَا هُمۡ يُنقَذُونَ ٤٣ إِلَّا رَحۡمَةٗ مِّنَّا وَمَتَٰعًا إِلَىٰ حِينٖ ٤٤ وَإِذَا قِيلَ لَهُمُ ٱتَّقُواْ مَا بَيۡنَ أَيۡدِيكُمۡ وَمَا خَلۡفَكُمۡ لَعَلَّكُمۡ تُرۡحَمُونَ ٤٥ وَمَا تَأۡتِيهِم مِّنۡ ءَايَةٖ مِّنۡ ءَايَٰتِ رَبِّهِمۡ إِلَّا كَانُواْ عَنۡهَا مُعۡرِضِينَ ٤٦ وَإِذَا قِيلَ لَهُمۡ أَنفِقُواْ مِمَّا رَزَقَكُمُ ٱللَّهُ قَالَ ٱلَّذِينَ كَفَرُواْ لِلَّذِينَ ءَامَنُوٓاْ أَنُطۡعِمُ مَن لَّوۡ يَشَآءُ ٱللَّهُ أَطۡعَمَهُۥٓ إِنۡ أَنتُمۡ إِلَّا فِي ضَلَٰلٖ مُّبِينٖ ٤٧ وَيَقُولُونَ مَتَىٰ هَٰذَا ٱلۡوَعۡدُ إِن كُنتُمۡ صَٰدِقِينَ ٤٨ مَا يَنظُرُونَ إِلَّا صَيۡحَةٗ وَٰحِدَةٗ تَأۡخُذُهُمۡ وَهُمۡ يَخِصِّمُونَ ٤٩ فَلَا يَسۡتَطِيعُونَ تَوۡصِيَةٗ وَلَآ إِلَىٰٓ أَهۡلِهِمۡ يَرۡجِعُونَ ٥٠ وَنُفِخَ فِي ٱلصُّورِ فَإِذَا هُم مِّنَ ٱلۡأَجۡدَاثِ إِلَىٰ رَبِّهِمۡ يَنسِلُونَ ٥١ قَالُواْ يَٰوَيۡلَنَا مَنۢ بَعَثَنَا مِن مَّرۡقَدِنَاۜۗ هَٰذَا مَا وَعَدَ ٱلرَّحۡمَٰنُ وَصَدَقَ ٱلۡمُرۡسَلُونَ ٥٢ إِن كَانَتۡ إِلَّا صَيۡحَةٗ وَٰحِدَةٗ فَإِذَا هُمۡ جَمِيعٞ لَّدَيۡنَا مُحۡضَرُونَ ٥٣ فَٱلۡيَوۡمَ لَا تُظۡلَمُ نَفۡسٞ شَيۡـٔٗا وَلَا تُجۡزَوۡنَ إِلَّا مَا كُنتُمۡ تَعۡمَلُونَ ٥٤ إِنَّ أَصۡحَٰبَ ٱلۡجَنَّةِ ٱلۡيَوۡمَ فِي شُغُلٖ فَٰكِهُونَ ٥٥ هُمۡ وَأَزۡوَٰجُهُمۡ فِي ظِلَٰلٍ عَلَى ٱلۡأَرَآئِكِ مُتَّكِـُٔونَ ٥٦ لَهُمۡ فِيهَا فَٰكِهَةٞ وَلَهُم مَّا يَدَّعُونَ ٥٧ سَلَٰمٞ قَوۡلٗا مِّن رَّبّٖ رَّحِيمٖ ٥٨ وَٱمۡتَٰزُواْ ٱلۡيَوۡمَ أَيُّهَا ٱلۡمُجۡرِمُونَ ٥٩ ۞أَلَمۡ أَعۡهَدۡ إِلَيۡكُمۡ يَٰبَنِيٓ ءَادَمَ أَن لَّا تَعۡبُدُواْ ٱلشَّيۡطَٰنَۖ إِنَّهُۥ لَكُمۡ عَدُوّٞ مُّبِينٞ ٦٠ وَأَنِ ٱعۡبُدُونِيۚ هَٰذَا صِرَٰطٞ مُّسۡتَقِيمٞ ٦١ وَلَقَدۡ أَضَلَّ مِنكُمۡ جِبِلّٗا كَثِيرًاۖ أَفَلَمۡ تَكُونُواْ تَعۡقِلُونَ ٦٢ هَٰذِهِۦ جَهَنَّمُ ٱلَّتِي كُنتُمۡ تُوعَدُونَ ٦٣ ٱصۡلَوۡهَا ٱلۡيَوۡمَ بِمَا كُنتُمۡ تَكۡفُرُونَ ٦٤ ٱلۡيَوۡمَ نَخۡتِمُ عَلَىٰٓ أَفۡوَٰهِهِمۡ وَتُكَلِّمُنَآ أَيۡدِيهِمۡ وَتَشۡهَدُ أَرۡجُلُهُم بِمَا كَانُواْ يَكۡسِبُونَ ٦٥ وَلَوۡ نَشَآءُ لَطَمَسۡنَا عَلَىٰٓ أَعۡيُنِهِمۡ فَٱسۡتَبَقُواْ ٱلصِّرَٰطَ فَأَنَّىٰ يُبۡصِرُونَ ٦٦ وَلَوۡ نَشَآءُ لَمَسَخۡنَٰهُمۡ عَلَىٰ مَكَانَتِهِمۡ فَمَا ٱسۡتَطَٰعُواْ مُضِيّٗا وَلَا يَرۡجِعُونَ ٦٧ وَمَن نُّعَمِّرۡهُ نُنَكِّسۡهُ فِي ٱلۡخَلۡقِۚ أَفَلَا يَعۡقِلُونَ ٦٨ وَمَا عَلَّمۡنَٰهُ ٱلشِّعۡرَ وَمَا يَنۢبَغِي لَهُۥٓۚ إِنۡ هُوَ إِلَّا ذِكۡرٞ وَقُرۡءَانٞ مُّبِينٞ ٦٩ لِّيُنذِرَ مَن كَانَ حَيّٗا وَيَحِقَّ ٱلۡقَوۡلُ عَلَى ٱلۡكَٰفِرِينَ ٧٠ أَوَلَمۡ يَرَوۡاْ أَنَّا خَلَقۡنَا لَهُم مِّمَّا عَمِلَتۡ أَيۡدِينَآ أَنۡعَٰمٗا فَهُمۡ لَهَا مَٰلِكُونَ ٧١ وَذَلَّلۡنَٰهَا لَهُمۡ فَمِنۡهَا رَكُوبُهُمۡ وَمِنۡهَا يَأۡكُلُونَ ٧٢ وَلَهُمۡ فِيهَا مَنَٰفِعُ وَمَشَارِبُۚ أَفَلَا يَشۡكُرُونَ ٧٣ وَٱتَّخَذُواْ مِن دُونِ ٱللَّهِ ءَالِهَةٗ لَّعَلَّهُمۡ يُنصَرُونَ ٧٤ لَا يَسۡتَطِيعُونَ نَصۡرَهُمۡ وَهُمۡ لَهُمۡ جُندٞ مُّحۡضَرُونَ ٧٥ فَلَا يَحۡزُنكَ قَوۡلُهُمۡۘ إِنَّا نَعۡلَمُ مَا يُسِرُّونَ وَمَا يُعۡلِنُونَ ٧٦ أَوَلَمۡ يَرَ ٱلۡإِنسَٰنُ أَنَّا خَلَقۡنَٰهُ مِن نُّطۡفَةٖ فَإِذَا هُوَ خَصِيمٞ مُّبِينٞ ٧٧ وَضَرَبَ لَنَا مَثَلٗا وَنَسِيَ خَلۡقَهُۥۖ قَالَ مَن يُحۡيِ ٱلۡعِظَٰمَ وَهِيَ رَمِيمٞ ٧٨ قُلۡ يُحۡيِيهَا ٱلَّذِيٓ أَنشَأَهَآ أَوَّلَ مَرَّةٖۖ وَهُوَ بِكُلِّ خَلۡقٍ عَلِيمٌ ٧٩ ٱلَّذِي جَعَلَ لَكُم مِّنَ ٱلشَّجَرِ ٱلۡأَخۡضَرِ نَارٗا فَإِذَآ أَنتُم مِّنۡهُ تُوقِدُونَ ٨٠ أَوَلَيۡسَ ٱلَّذِي خَلَقَ ٱلۡسَّمَٰوَٰتِ وَٱلۡأَرۡضَ بِقَٰدِرٍ عَلَىٰٓ أَن يَخۡلُقَ مِثۡلَهُمۚ بَلَىٰ وَهُوَ ٱلۡخَلَّٰقُ ٱلۡعَلِيمُ ٨١ إِنَّمَآ أَمۡرُهُۥٓ إِذَآ أَرَادَ شَيۡـًٔا أَن يَقُولَ لَهُۥ كُن فَيَكُونُ ٨٢ فَسُبۡحَٰنَ ٱلَّذِي بِيَدِهِۦ مَلَكُوتُ كُلِّ شَيۡءٖ وَإِلَيۡهِ تُرۡجَعُونَ ٨٣''',
+    translation: 'One time',
+    repeat: 1,
+  ),
+];
+
+///
+/// ──────────────────────────────────────────────────────────────────────────────
+/// MORNING & EVENING ADHKAR (Existing Code, Moved Here)
+/// ──────────────────────────────────────────────────────────────────────────────
+final List<DhikrItem> morningAdhkar = [
+  DhikrItem(
+    arabic:
+        'ٱللَّهُ لَآ إِلَٰهَ إِلَّا هُوَ ٱلۡحَيُّ '
+        'ٱلۡقَيُّومُۚ لَا تَأۡخُذُهُۥ سِنَةٞ وَلَا نَوۡمٞۚ '
+        'لَّهُۥ مَا فِي ٱلسَّمَٰوَٰتِ وَمَا فِي ٱلۡأَرۡضِۗ '
+        'مَن ذَا ٱلَّذِي يَشۡفَعُ عِندَهُۥٓ إِلَّا بِإِذۡنِهِۦۚ '
+        'يَعۡلَمُ مَا بَيۡنَ أَيۡدِيهِمۡ وَمَا خَلۡفَهُمۡۖ '
+        'وَلَا يُحِيطُونَ بِشَيۡءٖ مِّنۡ عِلۡمِهِۦٓ إِلَّا '
+        'بِمَا شَآءَۚ وَسِعَ كُرۡسِيُّهُ ٱلۡسَّمَٰوَٰتِ '
+        'وَٱلۡأَرۡضَۖ وَلَا يَـُٔودُهُۥ حِفۡظُهُمَاۚ '
+        'وَهُوَ ٱلۡعَلِيُّ ٱلۡعَظِيمُ ٢٥٥',
+    translation:
+        'Ayatul Kursi (Qur’an 2:255). “Allah – there is no deity except Him, '
+        'the Ever-Living, the Sustainer of [all] existence...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'بسم الله الرحمن الرحيم\n'
+        'قُلۡ هُوَ ٱللَّهُ أَحَدٌ ١ ٱللَّهُ '
+        'ٱلصَّمَدُ ٢ لَمۡ يَلِدۡ وَلَمۡ يُولَدۡ ٣ '
+        'وَلَمۡ يَكُن لَّهُۥ كُفُوًا أَحَدُۢ ٤',
+    translation:
+        'Surah Al-Ikhlas (Qur’an 112). “Say, He is Allah, [who is] One...”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic:
+        'بسم الله الرحمن الرحيم\n'
+        'قُلۡ أَعُوذُ بِرَبِّ ٱلۡفَلَقِ ١ '
+        'مِن شَرِّ مَا خَلَقَ ٢ وَمِن شَرِّ غَاسِقٍ '
+        'إِذَا وَقَبَ ٣ وَمِن شَرِّ ٱلنَّفَّٰثَٰتِ '
+        'فِي ٱلۡعُقَدِ ٤ وَمِن شَرِّ حَاسِدٍ '
+        'إِذَا حَسَدَ ٥',
+    translation:
+        'Surah Al-Falaq (Qur’an 113). “Say: I seek refuge in the Lord of daybreak...”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic:
+        'بسم الله الرحمن الرحيم\n'
+        'قُلۡ أَعُوذُ بِرَبِّ ٱلنَّاسِ ١ '
+        'مَلِكِ ٱلنَّاسِ ٢ إِلَٰهِ ٱلنَّاسِ ٣ '
+        'مِن شَرِّ ٱلۡوَسۡوَاسِ ٱلۡخَنَّاسِ ٤ '
+        'ٱلَّذِي يُوَسۡوِسُ فِي صُدُورِ ٱلنَّاسِ ٥ '
+        'مِنَ ٱلۡجِنَّةِ وَٱلنَّاسِ ٦',
+    translation:
+        'Surah An-Nas (Qur’an 114). “Say: I seek refuge in the Lord of mankind...”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic:
+        'أصبَحنا وَأصبَحَ المُلكُ للَّهِ، وَالحَمدُ للَّهِ، '
+        'لا إلَهَ إلّا اللَّهُ وَحدَهُ لا شَريكَ لَهُ، لَهُ '
+        'المُلكُ وَلَهُ الحَمدُ، وَهُوَ عَلى كُلِّ شَيءٍ '
+        'قَديرٌ، رَبِّ أسألُكَ خَيرَ ما في هَذا اليَومِ '
+        'وَخَيرَ ما بَعدَهُ، وَأعوذُ بِكَ مِن شَرِّ ما في '
+        'هَذا اليَومِ وَشَرِّ ما بَعدَهُ، رَبِّ أعوذُ بِكَ '
+        'مِنَ الكَسَلِ وَسُوءِ الكِبَرِ، رَبِّ أعوذُ بِكَ '
+        'مِن عَذابٍ في النّارِ وَعَذابٍ في القَبرِ',
+    translation:
+        '“We have reached the morning and at this very time all sovereignty '
+        'belongs to Allah... I seek refuge in You from the evil of this '
+        'day and what follows it...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ بِكَ أَصْبَحْنَا، وَبِكَ أَمْسَيْنَا، '
+        'وَبِكَ نَحْيَا، وَبِكَ نَمُوتُ وَإِلَيْكَ النُّشُورُ',
+    translation:
+        '“O Allah, by Your leave we have reached morning, and by Your leave '
+        'we have reached evening, by Your leave we live and die, and unto '
+        'You is the resurrection.”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ أَنْتَ رَبِّي لَا إِلَهَ إِلَّا أَنْتَ، '
+        'خَلَقْتَنِي وَأَنَا عَبْدُكَ، وَأَنَا عَلَى عَهْدِكَ '
+        'وَوَعْدِكَ مَا اسْتَطَعْتُ، أَعُوذُ بِكَ مِنْ شَرِّ '
+        'مَا صَنَعْتُ، أَبُوءُ لَكَ بِنِعْمَتِكَ عَلَيَّ، '
+        'وَأَبُوءُ بِذَنْبِي فَاغْفِرْ لِي فَإِنَّهُ لَا '
+        'يَغْفِرُ الذُّنُوبَ إِلَّا أَنْتَ',
+    translation:
+        '“O Allah, You are my Lord, none has the right to be worshiped but '
+        'You. You created me and I am Your servant... (Sayyid al-Istighfar).”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ إِنِّي أَصْبَحْتُ أُشْهِدُكَ، وَأُشْهِدُ '
+        'حَمَلَةَ عَرْشِكَ، وَمَلَائِكَتَكَ، وَجَمِيعَ خَلْقِكَ، '
+        'أَنَّكَ أَنْتَ اللَّهُ لَا إِلَهَ إِلَّا أَنْتَ '
+        'وَحْدَكَ لَا شَرِيكَ لَكَ، وَأَنَّ مُحَمَّدًا عَبْدُكَ '
+        'وَرَسُولُكَ',
+    translation:
+        '“O Allah, I have entered morning calling You to witness, and calling '
+        'the bearers of Your Throne, Your angels, and all Your creation to '
+        'witness that indeed You are Allah...”',
+    repeat: 4,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ مَا أَصْبَحَ بِي مِنْ نِعْمَةٍ أَوْ بِأَحَدٍ '
+        'مِنْ خَلْقِكَ فَمِنْكَ وَحْدَكَ لَا شَرِيكَ لَكَ، '
+        'فَلَكَ الْحَمْدُ وَلَكَ الشُّكْرُ',
+    translation:
+        '“O Allah, whatever blessing I or any of Your creation has this '
+        'morning is from You alone...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ عَافِنِي فِي بَدَنِي، اللَّهُمَّ عَافِنِي '
+        'فِي سَمْعِي، اللَّهُمَّ عَافِنِي فِي بَصَرِي, '
+        'لَا إِلَهَ إِلَّا أَنْتَ. اللَّهُمَّ إِنِّي '
+        'أَعُوذُ بِكَ مِنَ الْكُفْرِ، وَالْفَقْرِ، وَأَعُوذُ '
+        'بِكَ مِنْ عَذَابِ الْقَبْرِ، لَا إِلَهَ إِلَّا أَنْتَ',
+    translation:
+        '“O Allah, grant me well-being in my body, hearing, and sight... '
+        'I seek refuge in You from disbelief and poverty...”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic:
+        'حَسْبِيَ اللَّهُ لَا إِلَهَ إِلَّا هُوَ عَلَيْهِ '
+        'تَوَكَّلْتُ وَهُوَ رَبُّ الْعَرْشِ الْعَظِيمِ',
+    translation:
+        '“Allah is sufficient for me, there is no god but He; in Him I put '
+        'my trust...”',
+    repeat: 7,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ إِنِّي أَسْأَلُكَ الْعَفْوَ وَالْعَافِيَةَ '
+        'فِي الدُّنْيَا وَالْآخِرَةِ، اللَّهُمَّ إِنِّي '
+        'أَسْأَلُكَ الْعَفْوَ وَالْعَافِيَةَ فِي دِينِي '
+        'وَدُنْيَايَ وَأَهْلِي، وَمَالِي، اللَّهُمَّ اسْتُرْ '
+        'عَوْرَاتِي، وَآمِنْ رَوْعَاتِي، اللَّهُمَّ احْفَظْنِي '
+        'مِنْ بَيْنِ يَدَيَّ، وَمِنْ خَلْفِي، وَعَنْ يَمِينِي, '
+        'وَعَنْ شِمَالِي، وَمِنْ فَوْقِي، وَأَعُوذُ '
+        'بِعَظَمَتِكَ أَنْ أُغْتَالَ مِنْ تَحْتِي',
+    translation:
+        '“O Allah, I ask You for pardon and well-being in this life and the next... ”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ عَالِمَ الغَيْبِ وَالشَّهَادَةِ فَاطِرَ '
+        'السَّمَاوَاتِ وَالْأَرْضِ، رَبَّ كُلِّ شَيْءٍ '
+        'وَمَلِيكَهُ، أَشْهَدُ أَنْ لَا إِلَهَ إِلَّا أَنْتَ، '
+        'أَعُوذُ بِكَ مِنْ شَرِّ نَفْسِي، وَمِنْ شَرِّ '
+        'الشَّيْطَانِ وَشِرْكِهِ، وَأَنْ أَقْتَرِفَ عَلَى '
+        'نَفْسِي سُوءًا، أَوْ أَجُرَّهُ إِلَى مُسْلِمٍ',
+    translation:
+        '“O Allah, Knower of the unseen and the seen, Creator of the heavens '
+        'and earth, Lord of everything... I seek refuge in You from the evil '
+        'of myself and from Satan...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'بِسْمِ اللَّهِ الَّذِي لَا يَضُرُّ مَعَ اسْمِهِ '
+        'شَيْءٌ فِي الْأَرْضِ وَلَا فِي السَّمَاءِ وَهُوَ '
+        'السَّمِيعُ الْعَلِيمُ',
+    translation:
+        '“In the name of Allah, with whose name nothing in the earth or the '
+        'heavens can harm...”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic:
+        'رَضِيتُ بِاللَّهِ رَبًّا، وَبِالْإِسْلَامِ دِينًا، '
+        'وَبِمُحَمَّدٍ صلى الله عليه وسلم نَبِيًّا',
+    translation:
+        '“I am pleased with Allah as my Lord, Islam as my religion, and '
+        'Muhammad (peace be upon him) as my Prophet.”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic:
+        'يَا حَيُّ يَا قَيُّومُ بِرَحْمَتِكَ أَسْتَغيثُ '
+        'أَصْلِحْ لِي شَأْنِيَ كُلَّهُ وَلاَ تَكِلْنِي '
+        'إِلَى نَفْسِي طَرْفَةَ عَيْنٍ',
+    translation:
+        '“O Ever-Living, O Sustainer, by Your mercy I seek assistance, rectify '
+        'all of my affairs...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'أَصْبَحْنَا وَأَصْبَحَ الْمُلْكُ لِلَّهِ رَبِّ '
+        'الْعَالَمِينَ، اللَّهُمَّ إِنِّي أَسْأَلُكَ '
+        'خَيْرَ هَذَا الْيَوْمِ فَتْحَهُ، وَنَصْرَهُ، '
+        'وَنُورَهُ، وَبَرَكَتَهُ، وَهُدَاهُ، وَأَعُوذُ بِكَ '
+        'مِنْ شَرِّ مَا فِيهِ وَشَرِّ مَا بَعْدَهُ',
+    translation:
+        '“We have reached morning, and the entire kingdom belongs to Allah, '
+        'Lord of the worlds. O Allah, I ask You for the good of this day...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'أَصْبَحْنا عَلَى فِطْرَةِ الْإِسْلَامِ، وَعَلَى '
+        'كَلِمَةِ الْإِخْلَاصِ، وَعَلَى دِينِ نَبِيِّنَا '
+        'مُحَمَّدٍ صلى الله عليه وسلم، وَعَلَى مِلَّةِ '
+        'أَبِينَا إِبْرَاهِيمَ، حَنِيفاً مُسْلِماً وَمَا '
+        'كَانَ مِنَ الْمُشرِكِينَ',
+    translation:
+        '“We have awakened upon the natural religion of Islam, the word of '
+        'sincerity, the religion of our Prophet Muhammad...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: 'سُبْحَانَ اللَّهِ وَبِحَمْدِهِ',
+    translation: '“Glory be to Allah and praise be to Him.”',
+    repeat: 100,
+  ),
+  DhikrItem(
+    arabic:
+        'لاَ إِلَهَ إِلاَّ اللَّهُ وَحْدَهُ لاَ شَرِيكَ لَهُ، '
+        'لَهُ الْمُلْكُ وَلَهُ الْحَمْدُ، وَهُوَ عَلَى '
+        'كُلِّ شَيْءٍ قَدِيرٌ',
+    translation:
+        '“None has the right to be worshiped except Allah alone, without '
+        'partner. His is the dominion, and to Him is all praise...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'سُبْحَانَ اللَّهِ وَبِحَمْدِهِ, عَدَدَ خَلْقِهِ، '
+        'وَرِضَا نَفْسِهِ، وَزِنَةَ عَرْشِهِ، '
+        'وَمِدَادَ كَلِمَاتِهِ',
+    translation:
+        '“Glory be to Allah and praise be to Him, by the number of His creation, '
+        'His pleasure, the weight of His Throne...”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ إِنِّي أَسْأَلُكَ عِلْماً نَافِعاً، '
+        'وَرِزْقاً طَيِّباً، وَعَمَلاً مُتَقَبَّلاً',
+    translation:
+        '“O Allah, I ask You for beneficial knowledge, good provision, '
+        'and accepted deeds.”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: 'أَسْتَغْفِرُ اللَّهَ وَأَتُوبُ إِلَيْهِ',
+    translation: '“I seek Allah’s forgiveness and repent to Him.”',
+    repeat: 100,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ صَلِّ وَسَلِّمْ عَلَى نَبَيِّنَا مُحَمَّدٍ',
+    translation: '“O Allah, send Your blessings and peace upon our Prophet Muhammad.”',
+    repeat: 10,
+  ),
+];
+
+final List<DhikrItem> eveningAdhkar = [
+  DhikrItem(
+    arabic:
+        'ٱللَّهُ لَآ إِلَٰهَ إِلَّا هُوَ ٱلۡحَيُّ '
+        'ٱلۡقَيُّومُۚ لَا تَأۡخُذُهُۥ سِنَةٞ وَلَا نَوۡمٞۚ '
+        'لَّهُۥ مَا فِي ٱلۡسَّمَٰوَٰتِ وَمَا فِي ٱلۡأَرۡضِۗ '
+        'مَن ذَا ٱلَّذِي يَشۡفَعُ عِندَهُۥٓ إِلَّا '
+        'بِإِذۡنِهِۦۚ يَعۡلَمُ مَا بَيۡنَ أَيۡدِيهِمۡ '
+        'وَمَا خَلۡفَهُمۡۖ وَلَا يُحِيطُونَ بِشَيۡءٖ '
+        'مِّنۡ عِلۡمِهِۦٓ إِلَّا بِمَا شَآءَۚ وَسِعَ '
+        'كُرۡسِيُّهُ ٱلۡسَّمَٰوَٰتِ وَٱلۡأَرۡضَۖ '
+        'وَلَا يَـُٔودُهُۥ حِفۡظُهُمَاۚ وَهُوَ '
+        'ٱلۡعَلِيُّ ٱلۡعَظِيمُ ٢٥٥',
+    translation:
+        'Ayatul Kursi (Qur’an 2:255). Recite in the evening as well...',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'بسم الله الرحمن الرحيم\n'
+        'قُلۡ هُوَ ٱللَّهُ أَحَدٌ ١ '
+        'ٱللَّهُ ٱلصَّمَدُ ٢ لَمۡ يَلِدۡ '
+        'وَلَمۡ يُولَدۡ ٣ وَلَمۡ يَكُن لَّهُۥ '
+        'كُفُوًا أَحَدُۢ ٤',
+    translation:
+        'Surah Al-Ikhlas (Qur’an 112). “Say: He is Allah, the One...”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic:
+        'بسم الله الرحمن الرحيم\n'
+        'قُلۡ أَعُوذُ بِرَبِّ ٱلۡفَلَقِ ١ '
+        'مِن شَرِّ مَا خَلَقَ ٢ وَمِن شَرِّ '
+        'غَاسِقٍ إِذَا وَقَبَ ٣ وَمِن شَرِّ '
+        'ٱلنَّفَّٰثَٰتِ فِي ٱلۡعُقَدِ ٤ '
+        'وَمِن شَرِّ حَاسِدٍ إِذَا حَسَدَ ٥',
+    translation:
+        'Surah Al-Falaq (Qur’an 113). “Say: I seek refuge in the Lord of daybreak...”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic:
+        'بسم الله الرحمن الرحيم\n'
+        'قُلۡ أَعُوذُ بِرَبِّ ٱلنَّاسِ ١ '
+        'مَلِكِ ٱلنَّاسِ ٢ إِلَٰهِ '
+        'ٱلنَّاسِ ٣ مِن شَرِّ '
+        'ٱلۡوَسۡوَاسِ ٱلۡخَنَّاسِ ٤ '
+        'ٱلَّذِي يُوَسۡوِسُ فِي '
+        'صُدُورِ ٱلنَّاسِ ٥ مِنَ '
+        'ٱلۡجِنَّةِ وَٱلنَّاسِ ٦',
+    translation:
+        'Surah An-Nas (Qur’an 114). “Say: I seek refuge in the Lord of mankind...”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic:
+        'أمسَينا وَأمسى المُلكُ للَّهِ، وَالحَمدُ للَّهِ، '
+        'لا إلَهَ إلّا اللَّهُ وَحدَهُ لا شَريكَ لَهُ، '
+        'لَهُ المُلكُ وَلَهُ الحَمدُ، وَهُوَ عَلى كُلِّ '
+        'شَيءٍ قَديرٌ، رَبِّ أسألُكَ خَيرَ ما في هَذِهِ '
+        'اللَّيلَةِ وَخَيرَ ما بَعدَها، وَأعوذُ بِكَ '
+        'مِن شَرِّ ما في هَذِهِ اللَّيلَةِ وَشَرِّ '
+        'ما بَعدَها، رَبِّ أعوذُ بِكَ مِنَ الكَسَلِ '
+        'وَسُوءِ الكِبَرِ، رَبِّ أعوذُ بِكَ '
+        'مِن عَذابٍ في النّارِ وَعَذابٍ في القَبرِ',
+    translation:
+        '“We have entered the evening and at this very time all dominion belongs '
+        'to Allah... O Lord, I seek refuge in You from the evil of this night...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ بِكَ أمسَينا، وَبِكَ أصبَحنا، وَبِكَ '
+        'نَحيا، وَبِكَ نَموتُ، وَإلَيكَ المَصيرُ',
+    translation:
+        '“O Allah, by Your leave we have entered evening, and by Your leave we have '
+        'entered morning...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ أَنْتَ رَبِّي لَا إِلَهَ إِلَّا أَنْتَ، '
+        'خَلَقْتَنِي وَأَنَا عَبْدُكَ، وَأَنَا عَلَى عَهْدِكَ '
+        'وَوَعْدِكَ مَا اسْتَطَعْتُ، أَعُوذُ بِكَ مِنْ شَرِّ '
+        'مَا صَنَعْتُ، أَبُوءُ لَكَ بِنِعْمَتِكَ عَلَيَّ، '
+        'وَأَبُوءُ بِذَنْبِي فَاغْفِرْ لِي فَإِنَّهُ لَا '
+        'يَغْفِرُ الذُّنُوبَ إِلَّا أَنْتَ',
+    translation:
+        '“O Allah, You are my Lord, none has the right to be worshiped except You... ”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ إنّي أمسَيتُ أُشهِدُكَ، وَأُشهِدُ حَمَلَةَ '
+        'عَرشِكَ، وَمَلائِكَتَكَ، وَجَميعَ خَلقِكَ، أنَّكَ '
+        'أنتَ اللَّهُ لا إلَهَ إلّا أنتَ وَحدَكَ لا شَريكَ '
+        'لَكَ، وَأنَّ مُحَمَّدًا عَبدُكَ وَرَسولُكَ',
+    translation:
+        '“O Allah, I have entered upon the evening calling You to witness, and calling '
+        'the carriers of Your Throne, Your angels, and all of Your creation...”',
+    repeat: 4,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ ما أمسَى بي مِن نِعمَةٍ أو بِأحَدٍ مِن خَلقِكَ '
+        'فَمِنكَ وَحدَكَ لا شَريكَ لَكَ، فَلَكَ الحَمدُ '
+        'وَلَكَ الشُّكرُ',
+    translation:
+        '“O Allah, whatever blessing has come to me or any of Your creation this '
+        'evening, it is from You alone...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ عَافِنِي فِي بَدَنِي، اللَّهُمَّ عَافِنِي '
+        'فِي سَمْعِي، اللَّهُمَّ عَافِنِي فِي بَصَرِي، '
+        'لاَ إِلَهَ إِلاَّ أَنْتَ. اللَّهُمَّ إِنِّي '
+        'أَعُوذُ بِكَ مِنَ الْكُفْرِ، وَالفَقْرِ, '
+        'وَأَعُوذُ بِكَ مِنْ عَذَابِ القَبْرِ, '
+        'لاَ إِلَهَ إِلاَّ أَنْتَ',
+    translation:
+        '“O Allah, grant me health in my body, hearing, and sight... I seek refuge in '
+        'You from disbelief, poverty, and the torment of the grave...”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic:
+        'حَسْبِيَ اللَّهُ لاَ إِلَهَ إِلاَّ هُوَ عَلَيهِ '
+        'تَوَكَّلتُ وَهُوَ رَبُّ الْعَرْشِ الْعَظِيمِ',
+    translation:
+        '“Allah is sufficient for me; there is no deity but He. In Him I have placed '
+        'my trust, and He is the Lord of the Mighty Throne.”',
+    repeat: 7,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ إِنِّي أَسْأَلُكَ الْعَفْوَ وَالْعَافِيَةَ '
+        'فِي الدُّنْيَا وَالآخِرَةِ، اللَّهُمَّ إِنِّي '
+        'أَسْأَلُكَ الْعَفْوَ وَالْعَافِيَةَ فِي دِينِي '
+        'وَدُنْيَايَ وَأَهْلِي، وَمَالِي، اللَّهُمَّ '
+        'اسْتُرْ عَوْرَاتِي، وَآمِنْ رَوْعَاتِي, '
+        'اللَّهُمَّ احْفَظْنِي مِنْ بَينِ يَدَيَّ, '
+        'وَمِنْ خَلْفِي, وَعَنْ يَمِينِي, '
+        'وَعَنْ شِمَالِي, وَمِنْ فَوْقِي, '
+        'وَأَعُوذُ بِعَظَمَتِكَ أَنْ أُغْتَالَ مِنْ تَحْتِي',
+    translation:
+        '“O Allah, I ask You for pardon and well-being in this world and the Hereafter... ”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'اللَّهُمَّ عَالِمَ الغَيْبِ وَالشَّهَادَةِ فَاطِرَ '
+        'السَّمَاوَاتِ وَالْأَرْضِ، رَبَّ كُلِّ شَيْءٍ '
+        'وَمَلِيكَهُ، أَشْهَدُ أَنْ لَا إِلَهَ إِلَّا أَنْتَ, '
+        'أَعُوذُ بِكَ مِنْ شَرِّ نَفْسِي، وَمِنْ شَرِّ '
+        'الشَّيْطَانِ وَشِرْكِهِ، وَأَنْ أَقْتَرِفَ عَلَى '
+        'نَفْسِي سُوءًا، أَوْ أَجُرَّهُ إِلَى مُسْلِمٍ',
+    translation:
+        '“O Allah, Knower of the unseen and the seen, Creator of the heavens and earth... ”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'بِسْمِ اللَّهِ الَّذِي لاَ يَضُرُّ مَعَ اسْمِهِ '
+        'شَيْءٌ فِي الْأَرْضِ وَلاَ فِي السّمَاءِ '
+        'وَهُوَ السَّمِيعُ الْعَلِيمُ',
+    translation:
+        '“In the name of Allah with Whose name nothing can harm on earth or in heaven...”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic:
+        'رَضِيتُ بِاللَّهِ رَبَّاً، وَبِالْإِسْلاَمِ دِيناً, '
+        'وَبِمُحَمَّدٍ صلى الله عليه وسلم نَبِيّاً',
+    translation:
+        '“I am pleased with Allah as my Lord, Islam as my religion, and Muhammad (pbuh) '
+        'as my Prophet.”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic:
+        'يَا حَيُّ يَا قَيُّومُ بِرَحْمَتِكَ أَسْتَغيثُ '
+        'أَصْلِحْ لِي شَأْنِيَ كُلَّهُ وَلاَ تَكِلْنِي '
+        'إِلَى نَفْسِي طَرْفَةَ عَيْنٍ',
+    translation:
+        '“O Ever-Living, O Sustainer, by Your mercy I seek help; rectify all my affairs...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'أمسَينا وَأمسَى المُلكُ للَّهِ رَبِّ العالَمينَ, '
+        'اللَّهُمَّ إنّي أسألُكَ خَيرَ هَذِهِ اللَّيلَةِ: '
+        'فَتحَها، وَنَصرَها، وَنورَها، وَبَرَكَتَها, '
+        'وَهُداها، وَأعوذُ بِكَ مِن شَرِّ ما فيها '
+        'وَشَرِّ ما بَعدَها',
+    translation:
+        '“We have reached the evening, and all dominion belongs to Allah, Lord of the worlds... ”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic:
+        'أمسَينا عَلَى فِطْرَةِ الْإِسْلاَمِ، وَعَلَى '
+        'كَلِمَةِ الْإِخْلاَصِ، وَعَلَى دِينِ '
+        'نَبِيِّنَا مُحَمَّدٍ صلى الله عليه وسلم، '
+        'وَعَلَى مِلَّةِ أَبِينَا إِبْرَاهِيمَ، '
+        'حَنِيفاً مُسْلِماً وَمَا كَانَ مِنَ '
+        'الْمُشرِكِينَ',
+    translation:
+        '“We have entered upon the natural religion of Islam, the word of sincerity, '
+        'the religion of Prophet Muhammad, and the way of our father Ibrahim...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: 'سُبْحَانَ اللَّهِ وَبِحَمْدِهِ',
+    translation: '“Glory be to Allah and praise be to Him.”',
+    repeat: 100,
+  ),
+  DhikrItem(
+    arabic:
+        'لاَ إِلَهَ إِلاَّ اللَّهُ وَحْدَهُ لاَ شَرِيكَ '
+        'لَهُ، لَهُ الْمُلْكُ وَلَهُ الْحَمْدُ، '
+        'وَهُوَ عَلَى كُلِّ شَيْءٍ قَدِيرٌ',
+    translation:
+        '“None has the right to be worshiped except Allah alone, without partner. '
+        'His is the dominion, and to Him belongs the praise...”',
+    repeat: 1,
+  ),
+  DhikrItem(
+    arabic: 'أَعُوذُ بِكَلِمَاتِ اللَّهِ التَّامَّاتِ '
+        'مِنْ شَرِّ مَا خَلَقَ',
+    translation:
+        '“I seek refuge in the perfect words of Allah from the evil of what He created.”',
+    repeat: 3,
+  ),
+  DhikrItem(
+    arabic: 'اللَّهُمَّ صَلِّ وَسَلِّمْ عَلَى نَبَيِّنَا مُحَمَّدٍ',
+    translation:
+        '“O Allah, send blessings and peace upon our Prophet Muhammad.”',
+    repeat: 10,
+  ),
+];
