+++ lib/pages/azkar_page.dart
@@ -2,554 +2,10 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:percent_indicator/percent_indicator.dart';
 import 'package:confetti/confetti.dart';
+import 'package:prayer/models/azakdata.dart';
 
-///
-/// ──────────────────────────────────────────────────────────────────────────────
-///  DATA MODEL
-/// ──────────────────────────────────────────────────────────────────────────────
-///
-class DhikrItem {
-  final String arabic;       // Full Arabic text
-  final String translation;  // Short English translation
-  final int repeat;          // How many times
+// NEW IMPORT: we bring in DhikrItem and all the adhkar data from azakdata.dart
 
-  DhikrItem({
-    required this.arabic,
-    required this.translation,
-    required this.repeat,
-  });
-}
-
-///
-/// ──────────────────────────────────────────────────────────────────────────────
-///  MORNING ADHKAR (Full Copy-Paste + Short Translations)
-/// ──────────────────────────────────────────────────────────────────────────────
-///
-/// EXACT Arabic text you provided (no ellipses). Each item includes a short
-/// English meaning/translation (sourced from common references).
-///
-final List<DhikrItem> morningAdhkar = [
-  DhikrItem(
-    arabic:
-        'ٱللَّهُ لَآ إِلَٰهَ إِلَّا هُوَ ٱلۡحَيُّ '
-        'ٱلۡقَيُّومُۚ لَا تَأۡخُذُهُۥ سِنَةٞ وَلَا نَوۡمٞۚ '
-        'لَّهُۥ مَا فِي ٱلسَّمَٰوَٰتِ وَمَا فِي ٱلۡأَرۡضِۗ '
-        'مَن ذَا ٱلَّذِي يَشۡفَعُ عِندَهُۥٓ إِلَّا بِإِذۡنِهِۦۚ '
-        'يَعۡلَمُ مَا بَيۡنَ أَيۡدِيهِمۡ وَمَا خَلۡفَهُمۡۖ '
-        'وَلَا يُحِيطُونَ بِشَيۡءٖ مِّنۡ عِلۡمِهِۦٓ إِلَّا '
-        'بِمَا شَآءَۚ وَسِعَ كُرۡسِيُّهُ ٱلسَّمَٰوَٰتِ '
-        'وَٱلۡأَرۡضَۖ وَلَا يَـُٔودُهُۥ حِفۡظُهُمَاۚ '
-        'وَهُوَ ٱلۡعَلِيُّ ٱلۡعَظِيمُ ٢٥٥',
-    translation:
-        'Ayatul Kursi (Qur’an 2:255). “Allah – there is no deity except Him, '
-        'the Ever-Living, the Sustainer of [all] existence...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'بسم الله الرحمن الرحيم\n'
-        'قُلۡ هُوَ ٱللَّهُ أَحَدٌ ١ ٱللَّهُ '
-        'ٱلصَّمَدُ ٢ لَمۡ يَلِدۡ وَلَمۡ يُولَدۡ ٣ '
-        'وَلَمۡ يَكُن لَّهُۥ كُفُوًا أَحَدُۢ ٤',
-    translation:
-        'Surah Al-Ikhlas (Qur’an 112). “Say, He is Allah, [who is] One...”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic:
-        'بسم الله الرحمن الرحيم\n'
-        'قُلۡ أَعُوذُ بِرَبِّ ٱلۡفَلَقِ ١ '
-        'مِن شَرِّ مَا خَلَقَ ٢ وَمِن شَرِّ غَاسِقٍ '
-        'إِذَا وَقَبَ ٣ وَمِن شَرِّ ٱلنَّفَّٰثَٰتِ '
-        'فِي ٱلۡعُقَدِ ٤ وَمِن شَرِّ حَاسِدٍ '
-        'إِذَا حَسَدَ ٥',
-    translation:
-        'Surah Al-Falaq (Qur’an 113). “Say: I seek refuge in the Lord of daybreak...”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic:
-        'بسم الله الرحمن الرحيم\n'
-        'قُلۡ أَعُوذُ بِرَبِّ ٱلنَّاسِ ١ '
-        'مَلِكِ ٱلنَّاسِ ٢ إِلَٰهِ ٱلنَّاسِ ٣ '
-        'مِن شَرِّ ٱلۡوَسۡوَاسِ ٱلۡخَنَّاسِ ٤ '
-        'ٱلَّذِي يُوَسۡوِسُ فِي صُدُورِ ٱلنَّاسِ ٥ '
-        'مِنَ ٱلۡجِنَّةِ وَٱلنَّاسِ ٦',
-    translation:
-        'Surah An-Nas (Qur’an 114). “Say: I seek refuge in the Lord of mankind...”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic:
-        'أصبَحنا وَأصبَحَ المُلكُ للَّهِ، وَالحَمدُ للَّهِ، '
-        'لا إلَهَ إلّا اللَّهُ وَحدَهُ لا شَريكَ لَهُ، لَهُ '
-        'المُلكُ وَلَهُ الحَمدُ، وَهُوَ عَلى كُلِّ شَيءٍ '
-        'قَديرٌ، رَبِّ أسألُكَ خَيرَ ما في هَذا اليَومِ '
-        'وَخَيرَ ما بَعدَهُ، وَأعوذُ بِكَ مِن شَرِّ ما في '
-        'هَذا اليَومِ وَشَرِّ ما بَعدَهُ، رَبِّ أعوذُ بِكَ '
-        'مِنَ الكَسَلِ وَسُوءِ الكِبَرِ، رَبِّ أعوذُ بِكَ '
-        'مِن عَذابٍ في النّارِ وَعَذابٍ في القَبرِ',
-    translation:
-        '“We have reached the morning and at this very time all sovereignty '
-        'belongs to Allah... I seek refuge in You from the evil of this '
-        'day and what follows it...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ بِكَ أَصْبَحْنَا، وَبِكَ أَمْسَيْنَا، '
-        'وَبِكَ نَحْيَا، وَبِكَ نَمُوتُ وَإِلَيْكَ النُّشُورُ',
-    translation:
-        '“O Allah, by Your leave we have reached morning, and by Your leave '
-        'we have reached evening, by Your leave we live and die, and unto '
-        'You is the resurrection.”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ أَنْتَ رَبِّي لَا إِلَهَ إِلَّا أَنْتَ، '
-        'خَلَقْتَنِي وَأَنَا عَبْدُكَ، وَأَنَا عَلَى عَهْدِكَ '
-        'وَوَعْدِكَ مَا اسْتَطَعْتُ، أَعُوذُ بِكَ مِنْ شَرِّ '
-        'مَا صَنَعْتُ، أَبُوءُ لَكَ بِنِعْمَتِكَ عَلَيَّ، '
-        'وَأَبُوءُ بِذَنْبِي فَاغْفِرْ لِي فَإِنَّهُ لَا '
-        'يَغْفِرُ الذُّنُوبَ إِلَّا أَنْتَ',
-    translation:
-        '“O Allah, You are my Lord, none has the right to be worshiped but '
-        'You. You created me and I am Your servant... (Sayyid al-Istighfar).”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ إِنِّي أَصْبَحْتُ أُشْهِدُكَ، وَأُشْهِدُ '
-        'حَمَلَةَ عَرْشِكَ، وَمَلَائِكَتَكَ، وَجَمِيعَ خَلْقِكَ، '
-        'أَنَّكَ أَنْتَ اللَّهُ لَا إِلَهَ إِلَّا أَنْتَ '
-        'وَحْدَكَ لَا شَرِيكَ لَكَ، وَأَنَّ مُحَمَّدًا عَبْدُكَ '
-        'وَرَسُولُكَ',
-    translation:
-        '“O Allah, I have entered morning calling You to witness, and calling '
-        'the bearers of Your Throne, Your angels, and all Your creation to '
-        'witness that indeed You are Allah...”',
-    repeat: 4,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ مَا أَصْبَحَ بِي مِنْ نِعْمَةٍ أَوْ بِأَحَدٍ '
-        'مِنْ خَلْقِكَ فَمِنْكَ وَحْدَكَ لَا شَرِيكَ لَكَ، '
-        'فَلَكَ الْحَمْدُ وَلَكَ الشُّكْرُ',
-    translation:
-        '“O Allah, whatever blessing I or any of Your creation has this '
-        'morning is from You alone...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ عَافِنِي فِي بَدَنِي، اللَّهُمَّ عَافِنِي '
-        'فِي سَمْعِي، اللَّهُمَّ عَافِنِي فِي بَصَرِي، '
-        'لَا إِلَهَ إِلَّا أَنْتَ. اللَّهُمَّ إِنِّي '
-        'أَعُوذُ بِكَ مِنَ الْكُفْرِ، وَالْفَقْرِ، وَأَعُوذُ '
-        'بِكَ مِنْ عَذَابِ الْقَبْرِ، لَا إِلَهَ إِلَّا أَنْتَ',
-    translation:
-        '“O Allah, grant me well-being in my body, hearing, and sight... '
-        'I seek refuge in You from disbelief and poverty...”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic:
-        'حَسْبِيَ اللَّهُ لَا إِلَهَ إِلَّا هُوَ عَلَيْهِ '
-        'تَوَكَّلْتُ وَهُوَ رَبُّ الْعَرْشِ الْعَظِيمِ',
-    translation:
-        '“Allah is sufficient for me, there is no god but He; in Him I put '
-        'my trust...”',
-    repeat: 7,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ إِنِّي أَسْأَلُكَ الْعَفْوَ وَالْعَافِيَةَ '
-        'فِي الدُّنْيَا وَالْآخِرَةِ، اللَّهُمَّ إِنِّي '
-        'أَسْأَلُكَ الْعَفْوَ وَالْعَافِيَةَ فِي دِينِي '
-        'وَدُنْيَايَ وَأَهْلِي، وَمَالِي، اللَّهُمَّ اسْتُرْ '
-        'عَوْرَاتِي، وَآمِنْ رَوْعَاتِي، اللَّهُمَّ احْفَظْنِي '
-        'مِنْ بَيْنِ يَدَيَّ، وَمِنْ خَلْفِي، وَعَنْ يَمِينِي، '
-        'وَعَنْ شِمَالِي، وَمِنْ فَوْقِي، وَأَعُوذُ '
-        'بِعَظَمَتِكَ أَنْ أُغْتَالَ مِنْ تَحْتِي',
-    translation:
-        '“O Allah, I ask You for pardon and well-being in this life and the next... ”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ عَالِمَ الغَيْبِ وَالشَّهَادَةِ فَاطِرَ '
-        'السَّمَاوَاتِ وَالْأَرْضِ، رَبَّ كُلِّ شَيْءٍ '
-        'وَمَلِيكَهُ، أَشْهَدُ أَنْ لَا إِلَهَ إِلَّا أَنْتَ، '
-        'أَعُوذُ بِكَ مِنْ شَرِّ نَفْسِي، وَمِنْ شَرِّ '
-        'الشَّيْطَانِ وَشِرْكِهِ، وَأَنْ أَقْتَرِفَ عَلَى '
-        'نَفْسِي سُوءًا، أَوْ أَجُرَّهُ إِلَى مُسْلِمٍ',
-    translation:
-        '“O Allah, Knower of the unseen and the seen, Creator of the heavens '
-        'and earth, Lord of everything... I seek refuge in You from the evil '
-        'of myself and from Satan...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'بِسْمِ اللَّهِ الَّذِي لَا يَضُرُّ مَعَ اسْمِهِ '
-        'شَيْءٌ فِي الْأَرْضِ وَلَا فِي السَّمَاءِ وَهُوَ '
-        'السَّمِيعُ الْعَلِيمُ',
-    translation:
-        '“In the name of Allah, with whose name nothing in the earth or the '
-        'heavens can harm...”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic:
-        'رَضِيتُ بِاللَّهِ رَبًّا، وَبِالْإِسْلَامِ دِينًا، '
-        'وَبِمُحَمَّدٍ صلى الله عليه وسلم نَبِيًّا',
-    translation:
-        '“I am pleased with Allah as my Lord, Islam as my religion, and '
-        'Muhammad (peace be upon him) as my Prophet.”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic:
-        'يَا حَيُّ يَا قَيُّومُ بِرَحْمَتِكَ أَسْتَغيثُ '
-        'أَصْلِحْ لِي شَأْنِيَ كُلَّهُ وَلاَ تَكِلْنِي '
-        'إِلَى نَفْسِي طَرْفَةَ عَيْنٍ',
-    translation:
-        '“O Ever-Living, O Sustainer, by Your mercy I seek assistance, rectify '
-        'all of my affairs...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'أَصْبَحْنَا وَأَصْبَحَ الْمُلْكُ لِلَّهِ رَبِّ '
-        'الْعَالَمِينَ، اللَّهُمَّ إِنِّي أَسْأَلُكَ '
-        'خَيْرَ هَذَا الْيَوْمِ فَتْحَهُ، وَنَصْرَهُ، '
-        'وَنُورَهُ، وَبَرَكَتَهُ، وَهُدَاهُ، وَأَعُوذُ بِكَ '
-        'مِنْ شَرِّ مَا فِيهِ وَشَرِّ مَا بَعْدَهُ',
-    translation:
-        '“We have reached morning, and the entire kingdom belongs to Allah, '
-        'Lord of the worlds. O Allah, I ask You for the good of this day...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'أَصْبَحْنا عَلَى فِطْرَةِ الْإِسْلَامِ، وَعَلَى '
-        'كَلِمَةِ الْإِخْلَاصِ، وَعَلَى دِينِ نَبِيِّنَا '
-        'مُحَمَّدٍ صلى الله عليه وسلم، وَعَلَى مِلَّةِ '
-        'أَبِينَا إِبْرَاهِيمَ، حَنِيفاً مُسْلِماً وَمَا '
-        'كَانَ مِنَ الْمُشرِكِينَ',
-    translation:
-        '“We have awakened upon the natural religion of Islam, the word of '
-        'sincerity, the religion of our Prophet Muhammad...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic: 'سُبْحَانَ اللَّهِ وَبِحَمْدِهِ',
-    translation: '“Glory be to Allah and praise be to Him.”',
-    repeat: 100,
-  ),
-  DhikrItem(
-    arabic:
-        'لاَ إِلَهَ إِلاَّ اللَّهُ وَحْدَهُ لاَ شَرِيكَ لَهُ، '
-        'لَهُ الْمُلْكُ وَلَهُ الْحَمْدُ، وَهُوَ عَلَى '
-        'كُلِّ شَيْءٍ قَدِيرٌ',
-    translation:
-        '“None has the right to be worshiped except Allah alone, without '
-        'partner. His is the dominion, and to Him is all praise...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'سُبْحَانَ اللَّهِ وَبِحَمْدِهِ, عَدَدَ خَلْقِهِ، '
-        'وَرِضَا نَفْسِهِ، وَزِنَةَ عَرْشِهِ، '
-        'وَمِدَادَ كَلِمَاتِهِ',
-    translation:
-        '“Glory be to Allah and praise be to Him, by the number of His creation, '
-        'His pleasure, the weight of His Throne...”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ إِنِّي أَسْأَلُكَ عِلْماً نَافِعاً، '
-        'وَرِزْقاً طَيِّباً، وَعَمَلاً مُتَقَبَّلاً',
-    translation:
-        '“O Allah, I ask You for beneficial knowledge, good provision, '
-        'and accepted deeds.”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic: 'أَسْتَغْفِرُ اللَّهَ وَأَتُوبُ إِلَيْهِ',
-    translation: '“I seek Allah’s forgiveness and repent to Him.”',
-    repeat: 100,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ صَلِّ وَسَلِّمْ عَلَى نَبَيِّنَا مُحَمَّدٍ',
-    translation: '“O Allah, send Your blessings and peace upon our Prophet Muhammad.”',
-    repeat: 10,
-  ),
-];
-
-///
-/// ──────────────────────────────────────────────────────────────────────────────
-///  EVENING ADHKAR (Full Copy-Paste + Short Translations)
-/// ──────────────────────────────────────────────────────────────────────────────
-///
-final List<DhikrItem> eveningAdhkar = [
-  DhikrItem(
-    arabic:
-        'ٱللَّهُ لَآ إِلَٰهَ إِلَّا هُوَ ٱلۡحَيُّ '
-        'ٱلۡقَيُّومُۚ لَا تَأۡخُذُهُۥ سِنَةٞ وَلَا نَوۡمٞۚ '
-        'لَّهُۥ مَا فِي ٱلسَّمَٰوَٰتِ وَمَا فِي ٱلۡأَرۡضِۗ '
-        'مَن ذَا ٱلَّذِي يَشۡفَعُ عِندَهُۥٓ إِلَّا '
-        'بِإِذۡنِهِۦۚ يَعۡلَمُ مَا بَيۡنَ أَيۡدِيهِمۡ '
-        'وَمَا خَلۡفَهُمۡۖ وَلَا يُحِيطُونَ بِشَيۡءٖ '
-        'مِّنۡ عِلۡمِهِۦٓ إِلَّا بِمَا شَآءَۚ وَسِعَ '
-        'كُرۡسِيُّهُ ٱلسَّمَٰوَٰتِ وَٱلۡأَرۡضَۖ '
-        'وَلَا يَـُٔودُهُۥ حِفۡظُهُمَاۚ وَهُوَ '
-        'ٱلۡعَلِيُّ ٱلۡعَظِيمُ ٢٥٥',
-    translation:
-        'Ayatul Kursi (Qur’an 2:255). Recite in the evening as well...',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'بسم الله الرحمن الرحيم\n'
-        'قُلۡ هُوَ ٱللَّهُ أَحَدٌ ١ '
-        'ٱللَّهُ ٱلصَّمَدُ ٢ لَمۡ يَلِدۡ '
-        'وَلَمۡ يُولَدۡ ٣ وَلَمۡ يَكُن لَّهُۥ '
-        'كُفُوًا أَحَدُۢ ٤',
-    translation:
-        'Surah Al-Ikhlas (Qur’an 112). “Say: He is Allah, the One...”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic:
-        'بسم الله الرحمن الرحيم\n'
-        'قُلۡ أَعُوذُ بِرَبِّ ٱلۡفَلَقِ ١ '
-        'مِن شَرِّ مَا خَلَقَ ٢ وَمِن شَرِّ '
-        'غَاسِقٍ إِذَا وَقَبَ ٣ وَمِن شَرِّ '
-        'ٱلنَّفَّٰثَٰتِ فِي ٱلۡعُقَدِ ٤ '
-        'وَمِن شَرِّ حَاسِدٍ إِذَا حَسَدَ ٥',
-    translation:
-        'Surah Al-Falaq (Qur’an 113). “Say: I seek refuge in the Lord of daybreak...”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic:
-        'بسم الله الرحمن الرحيم\n'
-        'قُلۡ أَعُوذُ بِرَبِّ ٱلنَّاسِ ١ '
-        'مَلِكِ ٱلنَّاسِ ٢ إِلَٰهِ '
-        'ٱلنَّاسِ ٣ مِن شَرِّ '
-        'ٱلۡوَسۡوَاسِ ٱلۡخَنَّاسِ ٤ '
-        'ٱلَّذِي يُوَسۡوِسُ فِي '
-        'صُدُورِ ٱلنَّاسِ ٥ مِنَ '
-        'ٱلۡجِنَّةِ وَٱلنَّاسِ ٦',
-    translation:
-        'Surah An-Nas (Qur’an 114). “Say: I seek refuge in the Lord of mankind...”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic:
-        'أمسَينا وَأمسى المُلكُ للَّهِ، وَالحَمدُ للَّهِ، '
-        'لا إلَهَ إلّا اللَّهُ وَحدَهُ لا شَريكَ لَهُ، '
-        'لَهُ المُلكُ وَلَهُ الحَمدُ، وَهُوَ عَلى كُلِّ '
-        'شَيءٍ قَديرٌ، رَبِّ أسألُكَ خَيرَ ما في هَذِهِ '
-        'اللَّيلَةِ وَخَيرَ ما بَعدَها، وَأعوذُ بِكَ '
-        'مِن شَرِّ ما في هَذِهِ اللَّيلَةِ وَشَرِّ '
-        'ما بَعدَها، رَبِّ أعوذُ بِكَ مِنَ الكَسَلِ '
-        'وَسُوءِ الكِبَرِ، رَبِّ أعوذُ بِكَ '
-        'مِن عَذابٍ في النّارِ وَعَذابٍ في القَبرِ',
-    translation:
-        '“We have entered the evening and at this very time all dominion belongs '
-        'to Allah... O Lord, I seek refuge in You from the evil of this night...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ بِكَ أمسَينا، وَبِكَ أصبَحنا، وَبِكَ '
-        'نَحيا، وَبِكَ نَموتُ، وَإلَيكَ المَصيرُ',
-    translation:
-        '“O Allah, by Your leave we have entered evening, and by Your leave we have '
-        'entered morning...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ أَنْتَ رَبِّي لَا إِلَهَ إِلَّا أَنْتَ، '
-        'خَلَقْتَنِي وَأَنَا عَبْدُكَ، وَأَنَا عَلَى عَهْدِكَ '
-        'وَوَعْدِكَ مَا اسْتَطَعْتُ، أَعُوذُ بِكَ مِنْ شَرِّ '
-        'مَا صَنَعْتُ، أَبُوءُ لَكَ بِنِعْمَتِكَ عَلَيَّ، '
-        'وَأَبُوءُ بِذَنْبِي فَاغْفِرْ لِي فَإِنَّهُ لَا '
-        'يَغْفِرُ الذُّنُوبَ إِلَّا أَنْتَ',
-    translation:
-        '“O Allah, You are my Lord, none has the right to be worshiped except You... ”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ إنّي أمسَيتُ أُشهِدُكَ، وَأُشهِدُ حَمَلَةَ '
-        'عَرشِكَ، وَمَلائِكَتَكَ، وَجَميعَ خَلقِكَ، أنَّكَ '
-        'أنتَ اللَّهُ لا إلَهَ إلّا أنتَ وَحدَكَ لا شَريكَ '
-        'لَكَ، وَأنَّ مُحَمَّدًا عَبدُكَ وَرَسولُكَ',
-    translation:
-        '“O Allah, I have entered upon the evening calling You to witness, and calling '
-        'the carriers of Your Throne, Your angels, and all of Your creation...”',
-    repeat: 4,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ ما أمسَى بي مِن نِعمَةٍ أو بِأحَدٍ مِن خَلقِكَ '
-        'فَمِنكَ وَحدَكَ لا شَريكَ لَكَ، فَلَكَ الحَمدُ '
-        'وَلَكَ الشُّكرُ',
-    translation:
-        '“O Allah, whatever blessing has come to me or any of Your creation this '
-        'evening, it is from You alone...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ عَافِنِي فِي بَدَنِي، اللَّهُمَّ عَافِنِي '
-        'فِي سَمْعِي، اللَّهُمَّ عَافِنِي فِي بَصَرِي، '
-        'لاَ إِلَهَ إِلاَّ أَنْتَ. اللَّهُمَّ إِنِّي '
-        'أَعُوذُ بِكَ مِنَ الْكُفْرِ، وَالفَقْرِ، '
-        'وَأَعُوذُ بِكَ مِنْ عَذَابِ القَبْرِ، '
-        'لاَ إِلَهَ إِلاَّ أَنْتَ',
-    translation:
-        '“O Allah, grant me health in my body, hearing, and sight... I seek refuge in '
-        'You from disbelief, poverty, and the torment of the grave...”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic:
-        'حَسْبِيَ اللَّهُ لاَ إِلَهَ إِلاَّ هُوَ عَلَيهِ '
-        'تَوَكَّلتُ وَهُوَ رَبُّ الْعَرْشِ الْعَظِيمِ',
-    translation:
-        '“Allah is sufficient for me; there is no deity but He. In Him I have placed '
-        'my trust, and He is the Lord of the Mighty Throne.”',
-    repeat: 7,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ إِنِّي أَسْأَلُكَ الْعَفْوَ وَالْعَافِيَةَ '
-        'فِي الدُّنْيَا وَالآخِرَةِ، اللَّهُمَّ إِنِّي '
-        'أَسْأَلُكَ الْعَفْوَ وَالْعَافِيَةَ فِي دِينِي '
-        'وَدُنْيَايَ وَأَهْلِي، وَمَالِي، اللَّهُمَّ '
-        'اسْتُرْ عَوْرَاتِي، وَآمِنْ رَوْعَاتِي، '
-        'اللَّهُمَّ احْفَظْنِي مِنْ بَينِ يَدَيَّ، '
-        'وَمِنْ خَلْفِي، وَعَنْ يَمِينِي، '
-        'وَعَنْ شِمَالِي، وَمِنْ فَوْقِي، '
-        'وَأَعُوذُ بِعَظَمَتِكَ أَنْ أُغْتَالَ مِنْ تَحْتِي',
-    translation:
-        '“O Allah, I ask You for pardon and well-being in this world and the Hereafter... ”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'اللَّهُمَّ عَالِمَ الغَيْبِ وَالشَّهَادَةِ فَاطِرَ '
-        'السَّمَاوَاتِ وَالْأَرْضِ، رَبَّ كُلِّ شَيْءٍ '
-        'وَمَلِيكَهُ، أَشْهَدُ أَنْ لَا إِلَهَ إِلَّا أَنْتَ، '
-        'أَعُوذُ بِكَ مِنْ شَرِّ نَفْسِي، وَمِنْ شَرِّ '
-        'الشَّيْطَانِ وَشِرْكِهِ، وَأَنْ أَقْتَرِفَ عَلَى '
-        'نَفْسِي سُوءًا، أَوْ أَجُرَّهُ إِلَى مُسْلِمٍ',
-    translation:
-        '“O Allah, Knower of the unseen and the seen, Creator of the heavens and earth... ”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'بِسْمِ اللَّهِ الَّذِي لاَ يَضُرُّ مَعَ اسْمِهِ '
-        'شَيْءٌ فِي الْأَرْضِ وَلاَ فِي السّمَاءِ '
-        'وَهُوَ السَّمِيعُ الْعَلِيمُ',
-    translation:
-        '“In the name of Allah with Whose name nothing can harm on earth or in heaven...”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic:
-        'رَضِيتُ بِاللَّهِ رَبَّاً، وَبِالْإِسْلاَمِ دِيناً، '
-        'وَبِمُحَمَّدٍ صلى الله عليه وسلم نَبِيّاً',
-    translation:
-        '“I am pleased with Allah as my Lord, Islam as my religion, and Muhammad (pbuh) '
-        'as my Prophet.”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic:
-        'يَا حَيُّ يَا قَيُّومُ بِرَحْمَتِكَ أَسْتَغيثُ '
-        'أَصْلِحْ لِي شَأْنِيَ كُلَّهُ وَلاَ تَكِلْنِي '
-        'إِلَى نَفْسِي طَرْفَةَ عَيْنٍ',
-    translation:
-        '“O Ever-Living, O Sustainer, by Your mercy I seek help; rectify all my affairs...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'أمسَينا وَأمسَى المُلكُ للَّهِ رَبِّ العالَمينَ، '
-        'اللَّهُمَّ إنّي أسألُكَ خَيرَ هَذِهِ اللَّيلَةِ: '
-        'فَتحَها، وَنَصرَها، وَنورَها، وَبَرَكَتَها، '
-        'وَهُداها، وَأعوذُ بِكَ مِن شَرِّ ما فيها '
-        'وَشَرِّ ما بَعدَها',
-    translation:
-        '“We have reached the evening, and all dominion belongs to Allah, Lord of the worlds... ”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic:
-        'أمسَينا عَلَى فِطْرَةِ الْإِسْلاَمِ، وَعَلَى '
-        'كَلِمَةِ الْإِخْلاَصِ، وَعَلَى دِينِ '
-        'نَبِيِّنَا مُحَمَّدٍ صلى الله عليه وسلم، '
-        'وَعَلَى مِلَّةِ أَبِينَا إِبْرَاهِيمَ، '
-        'حَنِيفاً مُسْلِماً وَمَا كَانَ مِنَ '
-        'الْمُشرِكِينَ',
-    translation:
-        '“We have entered upon the natural religion of Islam, the word of sincerity, '
-        'the religion of Prophet Muhammad, and the way of our father Ibrahim...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic: 'سُبْحَانَ اللَّهِ وَبِحَمْدِهِ',
-    translation: '“Glory be to Allah and praise be to Him.”',
-    repeat: 100,
-  ),
-  DhikrItem(
-    arabic:
-        'لاَ إِلَهَ إِلاَّ اللَّهُ وَحْدَهُ لاَ شَرِيكَ '
-        'لَهُ، لَهُ الْمُلْكُ وَلَهُ الْحَمْدُ، '
-        'وَهُوَ عَلَى كُلِّ شَيْءٍ قَدِيرٌ',
-    translation:
-        '“None has the right to be worshiped except Allah alone, without partner. '
-        'His is the dominion, and to Him belongs the praise...”',
-    repeat: 1,
-  ),
-  DhikrItem(
-    arabic: 'أَعُوذُ بِكَلِمَاتِ اللَّهِ التَّامَّاتِ '
-        'مِنْ شَرِّ مَا خَلَقَ',
-    translation:
-        '“I seek refuge in the perfect words of Allah from the evil of what He created.”',
-    repeat: 3,
-  ),
-  DhikrItem(
-    arabic: 'اللَّهُمَّ صَلِّ وَسَلِّمْ عَلَى نَبَيِّنَا مُحَمَّدٍ',
-    translation:
-        '“O Allah, send blessings and peace upon our Prophet Muhammad.”',
-    repeat: 10,
-  ),
-];
-
-///
-/// ──────────────────────────────────────────────────────────────────────────────
-///  MAIN ADVANCED PAGE (Azkar & Tasbih)
-/// ──────────────────────────────────────────────────────────────────────────────
-///
-/// You can reference this in your main nav, e.g.:
-/// final _pages = [
-///   PrayerTimesPage(),
-///   AzkarAndTasbihAdvancedPage(),
-///   QiblaPage(),
-///   ...
-/// ];
-///
 class AzkarAndTasbihAdvancedPage extends StatefulWidget {
   const AzkarAndTasbihAdvancedPage({Key? key}) : super(key: key);
 
@@ -639,7 +95,7 @@ class _AzkarMenuPage extends StatelessWidget {
                 MaterialPageRoute(
                   builder: (_) => AzkarReadingPage(
                     title: 'Morning Azkar',
-                    items: morningAdhkar,
+                    items: morningAdhkar, // from azakdata.dart
                   ),
                 ),
               );
@@ -657,7 +113,99 @@ class _AzkarMenuPage extends StatelessWidget {
                 MaterialPageRoute(
                   builder: (_) => AzkarReadingPage(
                     title: 'Evening Azkar',
-                    items: eveningAdhkar,
+                    items: eveningAdhkar, // from azakdata.dart
+                  ),
+                ),
+              );
+            },
+          ),
+          const SizedBox(height: 20),
+          // Examples: You can link Sleep azkar, Waking up azkar, etc. similarly:
+          _AzkarCard(
+            title: 'Sleep Azkar',
+            subtitle: 'أذكار النوم',
+            color: Colors.teal,
+            icon: Icons.bed,
+            onTap: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (_) => AzkarReadingPage(
+                    title: 'Sleep Azkar',
+                    items: sleepAzkar, // from azakdata.dart
+                  ),
+                ),
+              );
+            },
+          ),
+          const SizedBox(height: 20),
+          _AzkarCard(
+            title: 'Waking Up Azkar',
+            subtitle: 'أذكار الاستيقاظ',
+            color: Colors.deepPurple,
+            icon: Icons.wb_sunny_outlined,
+            onTap: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (_) => AzkarReadingPage(
+                    title: 'Waking Up Azkar',
+                    items: wakingUpAzkar, // from azakdata.dart
+                  ),
+                ),
+              );
+            },
+          ),
+          const SizedBox(height: 20),
+          _AzkarCard(
+            title: 'After Prayers',
+            subtitle: 'أذكار بعد الصلاة',
+            color: Colors.indigo,
+            icon: Icons.done_all,
+            onTap: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (_) => AzkarReadingPage(
+                    title: 'After Prayers Azkar',
+                    items: afterPrayersAzkar, // from azakdata.dart
+                  ),
+                ),
+              );
+            },
+          ),
+          const SizedBox(height: 20),
+          // Surah Al-Mulk & Surah Yaseen (the separate big lists)
+          _AzkarCard(
+            title: 'Surah Al-Mulk',
+            subtitle: 'سورة الملك',
+            color: Colors.redAccent,
+            icon: Icons.book_outlined,
+            onTap: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (_) => AzkarReadingPage(
+                    title: 'Surah Al-Mulk',
+                    items: surahAlMulkAzkar, // from azakdata.dart
+                  ),
+                ),
+              );
+            },
+          ),
+          const SizedBox(height: 20),
+          _AzkarCard(
+            title: 'Surah Yaseen',
+            subtitle: 'سورة يس',
+            color: Colors.orange,
+            icon: Icons.book,
+            onTap: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (_) => AzkarReadingPage(
+                    title: 'Surah Yaseen',
+                    items: surahYaseenAzkar, // from azakdata.dart
                   ),
                 ),
               );
@@ -669,7 +217,7 @@ class _AzkarMenuPage extends StatelessWidget {
   }
 }
 
-/// A neat card for "Morning" / "Evening"
+/// A neat card for "Morning" / "Evening" / Sleep, etc.
 class _AzkarCard extends StatelessWidget {
   final String title;
   final String subtitle;
@@ -735,7 +283,7 @@ class _AzkarCard extends StatelessWidget {
 
 ///
 /// ──────────────────────────────────────────────────────────────────────────────
-///  2) TASBIH PAGE (ADVANCED) 
+///  2) TASBIH PAGE (ADVANCED)
 ///      - 4 counters: 1 global & 3 named
 /// ──────────────────────────────────────────────────────────────────────────────
 class TasbihAdvancedPage extends StatefulWidget {
@@ -843,7 +391,7 @@ class _TasbihAdvancedPageState extends State<TasbihAdvancedPage> {
                       lineWidth: 12.0,
                       animation: true,
                       animationDuration: 300,
-                      animateFromLastPercent: true, // so it doesn't reset to 0
+                      animateFromLastPercent: true, 
                       percent: mainFraction,
                       center: Text(
                         '$globalCount / $globalTarget',
@@ -904,8 +452,8 @@ class _TasbihAdvancedPageState extends State<TasbihAdvancedPage> {
                 icon: const Icon(Icons.refresh),
                 label: const Text('Reset All'),
                 style: ElevatedButton.styleFrom(
-                  padding:
-                      const EdgeInsets.symmetric(horizontal: 30, vertical: 12),
+                  padding: const EdgeInsets.symmetric(
+                      horizontal: 30, vertical: 12),
                 ),
               ),
               const SizedBox(height: 20),
@@ -1122,7 +670,9 @@ class _AzkarReadingPageState extends State<AzkarReadingPage> {
                             child: Container(
                               width: double.infinity,
                               padding: const EdgeInsets.symmetric(
-                                  horizontal: 20, vertical: 30),
+                                horizontal: 20, 
+                                vertical: 30,
+                              ),
                               decoration: BoxDecoration(
                                 borderRadius: BorderRadius.circular(20),
                                 gradient: LinearGradient(
@@ -1152,7 +702,7 @@ class _AzkarReadingPageState extends State<AzkarReadingPage> {
                                       ),
                                     ),
                                     const SizedBox(height: 12),
-                                    // English translation
+                                    // We put your "translation" or short note below
                                     Text(
                                       item.translation,
                                       textAlign: TextAlign.justify,
