+++ pubspec.yaml
@@ -50,6 +50,7 @@ dependencies:
 
   sensors_plus: any
   quran: ^1.0.0
+  share_plus: ^10.1.4
 
   
 dev_dependencies:
